% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{Argmax}
\alias{Argmax}
\title{Calibrated parameters}
\usage{
Argmax(stability, clustering = FALSE)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}} or
\code{\link{GraphicalModel}}.}

\item{clustering}{logical indicating whether indices of calibrated clustering
parameters (\code{clustering=TRUE}) or variable selection
(\code{clustering=FALSE}) should be extracted. This argument is only used
if \code{stability} is the output from \code{\link{Clustering}}.}
}
\value{
A matrix of parameter values. The first column (\code{lambda}])
  denotes the calibrated hyper-parameter of the underlying algorithm. The
  second column (\code{pi}) is the calibrated threshold in
  selection/co-membership proportions. In multi-block graphical modelling,
  rows correspond to different blocks.
}
\description{
Extracts calibrated parameter values in stability selection.
}
\examples{
\dontrun{

## Graphical modelling

# Data simulation
set.seed(1)
simul <- SimulateGraphical(pk = 20)

# Stability selection
stab <- GraphicalModel(xdata = simul$data)

# Extracting calibrated parameters
args <- Argmax(stab)


## Clustering

# Data simulation
set.seed(1)
simul <- SimulateClustering(
  n = c(10, 30, 15)
)

# Consensus clustering
stab <- Clustering(xdata = simul$data, Lambda = c(1.1, 1.5, 2))

# Extracting calibrated parameters for variable selection
Argmax(stab, clustering = FALSE)

# Extracting calibrated parameters for co-membership
Argmax(stab, clustering = TRUE)
}

}
\seealso{
\code{\link{VariableSelection}}, \code{\link{GraphicalModel}}

Other calibration functions: 
\code{\link{Adjacency}()},
\code{\link{ArgmaxId}()},
\code{\link{CalibrationPlot}()},
\code{\link{Clusters}()},
\code{\link{SelectedVariables}()},
\code{\link{SelectionProportions}()}
}
\concept{calibration functions}
