% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARI.R
\name{ARI}
\alias{ARI}
\title{Calculate the adjusted rand index}
\usage{
ARI(ground_true_clusters, pred_clusters)
}
\arguments{
\item{ground_true_clusters}{the ground-truth clusters}

\item{pred_clusters}{the predicted clusters}
}
\description{
This function is to calculate the performance of the algorithm by 5 metrics, including the Rand index, HA, MA, FM and Jaccard.
}
\examples{
finalmetrics = ARI(gtc, finalrowColor)

}
