% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{AverageEffect}
\alias{AverageEffect}
\title{Average coefficients conditionally on selection}
\usage{
AverageEffect(stability, lambda_id = NULL, side = "X", comp = 1)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}}.}

\item{lambda_id}{parameter ID with respect to the grid \code{Lambda}. If
\code{NULL}, average coefficients across the models run with the calibrated
parameter are returned.}

\item{side}{character string indicating if coefficients of the predictor
(\code{side="X"}) or outcome (\code{side="Y"}) coefficients should be
returned. Option \code{side="Y"} is only applicable to PLS models.}

\item{comp}{component ID. Only applicable to PLS models.}
}
\value{
A matrix of average coefficients across resampling iterations.
}
\description{
Extracts the average coefficients of the (calibrated) models conditionally on
selection across resampling iterations. This function can be applied to the
output of \code{\link{VariableSelection}}.
}
\examples{
\dontrun{

# Example with univariate outcome
set.seed(1)
simul <- SimulateRegression(n = 100, pk = 50, family = "gaussian")
stab <- VariableSelection(xdata = simul$xdata, ydata = simul$ydata, family = "gaussian")
av_betas <- AverageEffect(stab)
head(av_betas)

# Regression with multivariate outcomes
set.seed(1)
simul <- SimulateRegression(n = 100, pk = c(20, 30), family = "gaussian")
stab <- VariableSelection(xdata = simul$xdata, ydata = simul$ydata, family = "mgaussian")
av_betas <- AverageEffect(stab)
head(av_betas)
}
}
\seealso{
\code{\link{VariableSelection}}
}
