% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_score.R
\name{BinomialProbabilities}
\alias{BinomialProbabilities}
\title{Binomial probabilities for stability score}
\usage{
BinomialProbabilities(q, N, pi, K, n_cat = 3)
}
\arguments{
\item{q}{average number of features selected by the underlying algorithm.}

\item{N}{total number of features.}

\item{pi}{threshold in selection proportions.}

\item{K}{number of resampling iterations used to compute the selection proportions.}

\item{n_cat}{number of categories used to compute the stability score.
Possible values are 2 or 3.}
}
\value{
a list of probabilities for each of the 2 or 3 categories of
selection proportions.
}
\description{
Computes the probabilities of observing each category of
selection proportions under the assumption of a uniform selection
procedure.
}
