% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_block.R
\name{BlockDiagonal}
\alias{BlockDiagonal}
\title{Block diagonal matrix}
\usage{
BlockDiagonal(pk)
}
\arguments{
\item{pk}{vector encoding the grouping structure.}
}
\value{
A binary block diagonal matrix.
}
\description{
Generates a binary block diagonal matrix.
}
\examples{
# Small example
mat <- BlockDiagonal(pk = c(2, 3))
}
