% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_block.R
\name{BlockMatrix}
\alias{BlockMatrix}
\title{Block matrix}
\usage{
BlockMatrix(pk)
}
\arguments{
\item{pk}{vector encoding the grouping structure.}
}
\value{
a symmetric block matrix.
}
\description{
Generates a symmetric matrix of the size of the adjacency matrix encoding the
block structure from the numbers of variables in each group.
}
\examples{
# Small example
mat <- BlockMatrix(pk = c(2, 3))
dim(mat)

}
\seealso{
Other multi-block functions: 
\code{\link{BlockLambdaGrid}()},
\code{\link{BlockStructure}()}
}
\concept{multi-block functions}
