% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{CalibrationCurve}
\alias{CalibrationCurve}
\title{Calibration curve (internal)}
\usage{
CalibrationCurve(
  stability,
  bty = "o",
  xlab = expression(n[C]),
  ylab = "Consensus score",
  col = c("navy", "forestgreen", "tomato"),
  legend = TRUE,
  ncol = 1
)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}},
\code{\link{GraphicalModel}} or \code{\link{BiSelection}}.}

\item{bty}{character string indicating if the box around the plot should be
drawn. Possible values include: \code{"o"} (default, the box is drawn), or
\code{"n"} (no box).}

\item{xlab}{label of the x-axis.}

\item{ylab}{label of the y-axis.}

\item{col}{vector of colours.}

\item{legend}{logical indicating if the legend should be included.}

\item{ncol}{integer indicating the number of columns in the legend.}
}
\value{
a calibration curve.
}
\description{
Creates a calibration curve for consensus clustering.
}
\keyword{internal}
