% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percluster_exp.R
\name{cfun}
\alias{cfun}
\title{Combine results which may contain NULL or sparse matrices}
\usage{
cfun(...)
}
\arguments{
\item{...}{variable number of inputs which are either sparse matrices or NULL}
}
\description{
Because by default, combining of sparse matrices and NULL will yield NULL which is not consistent with our expectation, we need to remove NULL to do the combining the sparse matrices. This function acts as a self-crafted combination function for the foreach output.
}
