% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{CheckPackageInstalled}
\alias{CheckPackageInstalled}
\title{Checking that a package is installed}
\usage{
CheckPackageInstalled(package)
}
\arguments{
\item{package}{character string indicating the name of the package.}
}
\description{
Checks if a package is installed and returns an error message if not.
}
\keyword{internal}
