% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_models.R
\name{CLARAClustering}
\alias{CLARAClustering}
\title{Clustering Large Applications}
\usage{
CLARAClustering(xdata, nc = NULL, scale = TRUE, rows = TRUE, ...)
}
\arguments{
\item{xdata}{data matrix with observations as rows and variables as columns.}

\item{nc}{matrix of parameters controlling the number of clusters. If
\code{nc} is not provided, it is set to \code{seq(1, nrow(xdata))}.}

\item{scale}{logical indicating if the data should be scaled to ensure that
all variables contribute equally to the clustering of the observations.}

\item{rows}{logical indicating if clusters of rows (\code{TRUE}) or columns
(\code{FALSE}) should be inferred.}

\item{...}{additional parameters passed to \code{\link[cluster]{clara}}.}
}
\value{
An array of binary and symmetric co-membership matrices.
}
\description{
Runs Clustering Large Applications (CLARA) algorithm using implementation
from \code{\link[cluster]{clara}}. This function applies the Partitioning
Around Medoids (PAM) algorithm on subsamples of the features to cluster. It
makes it more efficient for high-dimensional datasets. This function is not
using stability.
}
\examples{

# Data simulation
set.seed(1)
simul <- SimulateClustering(n = c(10, 10), pk = 50)

# PAM clustering
mypam <- CLARAClustering(xdata = simul$data, nc = 1:20)
}
\seealso{
Other clustering algorithms: 
\code{\link{DBSCANClustering}()},
\code{\link{GMMClustering}()},
\code{\link{HierarchicalClustering}()},
\code{\link{KMeansClustering}()},
\code{\link{PAMClustering}()}
}
\concept{clustering algorithms}
