% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{Coefficients}
\alias{Coefficients}
\title{Model coefficients}
\usage{
Coefficients(stability, side = "X", comp = 1, iterations = NULL)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}}.}

\item{side}{character string indicating if coefficients of the predictor
(\code{side="X"}) or outcome (\code{side="Y"}) coefficients should be
returned. Option \code{side="Y"} is only applicable to PLS models.}

\item{comp}{component ID. Only applicable to PLS models.}

\item{iterations}{vector of iteration IDs. If \code{iterations=NULL}, the
coefficients of all visited models are returned. This number must be
smaller than the number of iterations \code{K} used for the stability
selection run.}
}
\description{
Extracts the coefficients of visited models at different resampling
iterations of a stability selection run. This function can be applied to the
output of \code{\link{VariableSelection}}.
}
\examples{
\dontrun{

# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 100, pk = 50, family = "gaussian")

# Stability selection
stab <- VariableSelection(xdata = simul$xdata, ydata = simul$ydata, family = "gaussian")

# Coefficients of visited models
coefs <- focus:::Coefficients(stab)
dim(coefs)

# Coefficients of the first fitted model
coefs <- focus:::Coefficients(stab, iterations = 1)
dim(coefs)

# Stability selection
stab <- VariableSelection(
  xdata = simul$xdata, ydata = simul$ydata,
  implementation = SparsePLS, family = "gaussian"
)

# Coefficients of visited models
coefs <- focus:::Coefficients(stab, side = "Y", )
dim(coefs)
}

}
\seealso{
\code{\link{VariableSelection}}
}
\keyword{internal}
