% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CoMembership}
\alias{CoMembership}
\title{Pairwise co-membership}
\usage{
CoMembership(groups)
}
\arguments{
\item{groups}{vector of group membership.}
}
\value{
A symmetric and binary matrix.
}
\description{
Generates a symmetric and binary matrix indicating if a pair of features
belongs to the same cluster.
}
