% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Concatenate}
\alias{Concatenate}
\title{Concatenate stability objects}
\usage{
Concatenate(stability1, stability2 = NULL, order_output = FALSE)
}
\arguments{
\item{stability1}{a stability object.}

\item{stability2}{another stability object (optional).}
}
\value{
A single stability object.
}
\description{
Generates a single stability object from two stability objects. This function
is used to concatenate results when using \code{\link[nloptr]{nloptr}}.
}
\keyword{internal}
