% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConcatenateVariableSelection}
\alias{ConcatenateVariableSelection}
\title{Concatenate stability objects for variable selection}
\usage{
ConcatenateVariableSelection(
  stability1,
  stability2 = NULL,
  order_output = FALSE
)
}
\arguments{
\item{stability1}{matrix of dummy variables.}

\item{stability2}{logical indicating if messages should be printed.}
}
\value{
A single stability object.
}
\description{
Generates a single stability object from two stability objects. This function
is used to concatenate results when using nloptr for optimisation in
\code{\link{VariableSelection}}.
}
\keyword{internal}
