% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{ConsensusMatrix}
\alias{ConsensusMatrix}
\title{Calibrated consensus matrix}
\usage{
ConsensusMatrix(stability, argmax_id = NULL)
}
\arguments{
\item{stability}{output of \code{\link{Clustering}}.}

\item{argmax_id}{optional matrix of parameter IDs. If \code{argmax_id=NULL},
the calibrated model is used.}
}
\value{
A consensus matrix.
}
\description{
Extracts the (calibrated) consensus matrix. Entries indicate the proportion
of subsampling iterations where the two items were in the same cluster, out
of all iterations where both items were drawn in the subsample.
}
\examples{
\dontrun{

# Simulation of data with clusters
set.seed(1)
simul <- SimulateClustering(
  n = c(10, 30, 15),
  pk = 10, nu_xc = 1,
  ev_xc = 0.95
)

# Consensus clustering
stab <- Clustering(xdata = x, implementation = HierarchicalClustering)
ConsensusMatrix(stab)
}

}
\seealso{
\code{\link{Clustering}}

Other calibration functions: 
\code{\link{ArgmaxId}()},
\code{\link{SelectionProportions}()},
\code{\link{Stable}()}
}
\concept{calibration functions}
