% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{Contrast}
\alias{Contrast}
\title{Contrast of a matrix}
\usage{
Contrast(mat, digits = 3)
}
\arguments{
\item{mat}{input matrix for which to compute the contrast.}

\item{digits}{number of digits to use.}
}
\value{
A single number, the contrast of the input matrix.
}
\description{
Computes the contrast of a matrix, defined as the number of visited bins of entries
with a specified number of digits.
}
