% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus.R
\docType{package}
\name{focus}
\alias{focus}
\title{focus: Feature selectiOn and Clustering Using Stability}
\description{
Implementation of stability-enhanced models for variable selection in
regression, network estimation and clustering. These methods are based on
resampling approaches to compute selection proportions. Calibration of the
models is done via maximisation of a stability score measuring how unlikely
it is that the selection procedure is uniform.
}
