% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percluster_exp.R
\name{get_percluster_exp}
\alias{get_percluster_exp}
\title{Get cluster-specific expression matrix}
\usage{
get_percluster_exp(scExp, y, n.cores)
}
\arguments{
\item{scExp}{the list of original block-wise expression matrices}

\item{y}{the clustering results by running SHARP}

\item{n.cores}{number of cores to be used. The default is (n-1) cores, where n is the number of cores in your local computer or server.}
}
\value{
a list of sparse matrices where each matrix corresponds to the expression matrix of each cluster
}
\description{
This is to get the cluster-specific expresion matrix for further detecting cluster-specific marker genes, given the original scRNA-seq data (in the list format) and the clustering results.
}
\examples{

y = SHARP_unlimited(scExp)
smat = get_percluster_exp(scExp, y)


}
