% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMetaC.R
\name{getA}
\alias{getA}
\title{Obtain the co-location matrix}
\usage{
getA(rowColor)
}
\arguments{
\item{rowColor}{the clustering results by some clustering algorithms}
}
\description{
This function is to obtain the weighted co-location matrix from the clustering result for following ensemble clustering.
}
\examples{
AA = getA(rowColor)

}
