% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrowColor.R
\name{gethclust}
\alias{gethclust}
\title{hierarchical clustering with number of clusters determined automatically}
\usage{
gethclust(d, my)
}
\description{
This function is to do hierarchical clustering, with the number of clusters determined by a strategy combining Silhouette index, CH index and the heights after hierarchical clustering.
}
