% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrowColor.R
\name{getrowColor}
\alias{getrowColor}
\title{hierarchical clustering with number of clusters determined automatically}
\usage{
getrowColor(Emat, indN.cluster)
}
\arguments{
\item{Emat}{the feature matrix whose columns represent the features and whose rows represent data/cells.}
}
\description{
This function is to do hierarchical clustering, with the number of clusters determined by a strategy combining Silhouette index, CH index and the heights after hierarchical clustering.
}
\examples{
rowColor= getrowColor(E1)

}
