% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_models.R
\name{HierarchicalClustering}
\alias{HierarchicalClustering}
\title{Hierarchical clustering}
\usage{
HierarchicalClustering(xdata, Lambda = NULL, scale = TRUE, ...)
}
\arguments{
\item{xdata}{data matrix with observations as rows and variables as columns.}

\item{Lambda}{matrix of parameters controlling the number of clusters in the
underlying algorithm specified in \code{implementation}. If \code{Lambda}
is not provided, it is set to \code{seq(1, nrow(xdata))}.}

\item{scale}{logical indicating if the data should be scaled to ensure that
all variables contribute equally to the clustering of the observations.}

\item{...}{additional parameters passed to \code{\link[stats]{hclust}}.}
}
\value{
An array with binary and symmetric co-membership matrices.
}
\description{
Runs hierarchical clustering using implementation from
\code{\link[stats]{hclust}}. This function is not using stability.
}
\examples{

# Data simulation
set.seed(1)
simul <- SimulateRegression(pk = 50, n = 10)

myhclust <- HierarchicalClustering(xdata = simul$X, Lambda = 1:5)
}
