% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{KMeansClustering}
\alias{KMeansClustering}
\title{(Sparse) K-means clustering}
\usage{
KMeansClustering(xdata, nc = NULL, Lambda = NULL, ...)
}
\arguments{
\item{xdata}{data matrix with observations as rows and variables as columns.}

\item{nc}{matrix of parameters controlling the number of clusters in the
underlying algorithm specified in \code{implementation}. If \code{nc} is
not provided, it is set to \code{seq(1, tau*nrow(xdata))}.}

\item{Lambda}{vector of penalty parameters (see argument \code{wbounds} in
\code{\link[sparcl]{KMeansSparseCluster}}).}

\item{...}{additional parameters passed to \code{\link[stats]{kmeans}} (if
\code{Lambda} is \code{NULL}) or \code{\link[sparcl]{KMeansSparseCluster}}.}
}
\value{
A list with: \item{comembership}{an array of binary and symmetric
  co-membership matrices.} \item{weights}{a matrix of median weights by
  feature.}
}
\description{
Runs k-means clustering using implementation from
\code{\link[stats]{kmeans}}. This function is not using stability.
}
\examples{

# Data simulation
set.seed(1)
simul <- SimulateClustering(n = c(10, 10), pk = 50)

# K means clustering
mykmeans <- KMeansClustering(xdata = simul$data, nc = seq_len(20))

# Sparse K means clustering
if (requireNamespace("sparcl", quietly = TRUE)) {
  mykmeans <- KMeansClustering(
    xdata = simul$data, nc = seq_len(20),
    Lambda = c(2, 5)
  )
}

}
\references{
\insertRef{SparseClustering}{sharp}
}
\seealso{
Other clustering algorithms: 
\code{\link{DBSCANClustering}()},
\code{\link{GMMClustering}()},
\code{\link{HierarchicalClustering}()},
\code{\link{PAMClustering}()}
}
\concept{clustering algorithms}
