% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_models.R
\name{LavaanMatrix}
\alias{LavaanMatrix}
\title{Matrix from lavaan outputs}
\usage{
LavaanMatrix(vect, adjacency, residual_covariance = NULL)
}
\arguments{
\item{vect}{vector of coefficients to assign to entries of the matrix.}

\item{adjacency}{binary adjacency matrix of the Directed Acyclic Graph
(asymmetric matrix A in Reticular Action Model notation).}

\item{residual_covariance}{binary and symmetric matrix encoding the nonzero
entries in the residual covariance matrix (symmetric matrix S in Reticular
Action Model notation). By default, this is the identity matrix (no
residual covariance).}
}
\value{
An asymmetric matrix.
}
\description{
Returns a matrix from output in \code{\link[lavaan]{lavaan}} format.
}
\examples{
# Definition of the model structure
layers <- list(
  c("var1", "var2", "var3"),
  c("var4", "var5"),
  c("var6", "var7", "var8")
)
dag <- LayeredDAG(layers)

# Definition of simulated effects
theta <- dag
theta[2, 4] <- 0

# Data simulation
set.seed(1)
simul <- SimulateStructural(theta = theta)

# Running regularised SEM
mysem <- PenalisedSEM(
  xdata = simul$data, adjacency = dag,
  Lambda = 0.3
)

# Obtaining results in matrix format
LavaanMatrix(vect = mysem$beta_full, adjacency = dag)

}
\references{
\insertRef{lavaanBook}{sharp}
}
\seealso{
\code{\link{LavaanModel}}, \code{\link{PenalisedSEM}}
}
