% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_models.R
\name{LinearSystemMatrix}
\alias{LinearSystemMatrix}
\title{Matrix from linear system outputs}
\usage{
LinearSystemMatrix(vect, adjacency)
}
\arguments{
\item{vect}{vector of coefficients to assign to entries of the matrix.}

\item{adjacency}{binary adjacency matrix of the Directed Acyclic Graph
(transpose of the asymmetric matrix A in Reticular Action Model notation).
The row and column names of this matrix must be defined.}
}
\value{
An asymmetric matrix.
}
\description{
Returns a matrix from output of \code{\link{PenalisedLinearSystem}}.
}
\seealso{
\code{\link{PenalisedLinearSystem}}
}
