% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{MaxContrast}
\alias{MaxContrast}
\title{Maximising matrix contrast}
\usage{
MaxContrast(u, omega, digits = 3)
}
\arguments{
\item{u}{constant u added to the diagonal of the precision matrix.}

\item{omega}{positive semi-definite precision matrix.}

\item{digits}{number of digits to use in the definition of the contrast.}
}
\value{
A single number, the contrast of the generated precision matrix.
}
\description{
Computes the contrast of the correlation matrix obtained by adding u to the
diagonal of the precision matrix. This function is used to find the value of
u that maximises the contrast when constructing a diagonally dominant
precision matrix.
}
