% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NAToNULL}
\alias{NAToNULL}
\title{Transforms NA into NULL}
\usage{
NAToNULL(x)
}
\arguments{
\item{x}{input vector.}
}
\value{
a vector without missing values or NULL.
}
\description{
Returns a vector with no missing values or NULL if there are no non-missing
values.
}
