% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_models.R
\name{OpenMxModel}
\alias{OpenMxModel}
\title{Writing OpenMx model (matrix specification)}
\usage{
OpenMxModel(adjacency, residual_covariance = NULL)
}
\arguments{
\item{adjacency}{binary adjacency matrix of the Directed Acyclic Graph
(transpose of the asymmetric matrix A in Reticular Action Model notation).}

\item{residual_covariance}{binary and symmetric matrix encoding the nonzero
entries in the residual covariance matrix (symmetric matrix S in Reticular
Action Model notation). By default, this is the identity matrix (no
residual covariance).}
}
\value{
A character string that can be used in argument \code{model} in
  \code{\link[lavaan]{sem}}.
}
\description{
Returns matrix specification for use in \code{\link[OpenMx]{mxModel}} from
(i) the adjacency matrix of a Directed Acyclic Graph (asymmetric matrix A in
Reticular Action Model notation), and (ii) a binary matrix encoding nonzero
entries in the residual covariance matrix (symmetric matrix S in Reticular
Action Model notation).
}
\examples{
# Definition of the model structure
layers <- list(
  c("var1", "var2", "var3"),
  c("var4", "var5"),
  c("var6", "var7", "var8")
)
dag <- LayeredDAG(layers)

# Definition of simulated effects
theta <- dag
theta[2, 4] <- 0
theta[3, 7] <- 0
theta[4, 7] <- 0

# Data simulation
set.seed(1)
simul <- SimulateStructural(n = 500, v_between = 1, theta = theta)

# Writing RAM matrices for mxModel
ram_matrices <- OpenMxModel(adjacency = dag)

# Running unpenalised model
unpenalised <- OpenMx::mxRun(OpenMx::mxModel(
  "Model",
  OpenMx::mxData(simul$data, type = "raw"),
  ram_matrices$Amat,
  ram_matrices$Smat,
  ram_matrices$Fmat,
  ram_matrices$Mmat,
  OpenMx::mxExpectationRAM("A", "S", "F", "M", dimnames = colnames(dag)),
  OpenMx::mxFitFunctionML()
), silent = TRUE, suppressWarnings = TRUE)

}
\references{
\insertRef{lavaanBook}{sharp}
}
\seealso{
\code{\link{PenalisedSEM}}, \code{\link{LavaanMatrix}}
}
