% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_models.R
\name{PAMClustering}
\alias{PAMClustering}
\title{Partitioning Around Medoids}
\usage{
PAMClustering(xdata, nc = NULL, scale = TRUE, rows = TRUE, ...)
}
\arguments{
\item{xdata}{data matrix with observations as rows and variables as columns.}

\item{nc}{matrix of parameters controlling the number of clusters. If
\code{nc} is not provided, it is set to \code{seq(1, nrow(xdata))}.}

\item{scale}{logical indicating if the data should be scaled to ensure that
all variables contribute equally to the clustering of the observations.}

\item{rows}{logical indicating if clusters of rows (\code{TRUE}) or columns
(\code{FALSE}) should be inferred.}

\item{...}{additional parameters passed to \code{\link[stats]{dist}} or
\code{\link[cluster]{pam}}.}
}
\value{
An array of binary and symmetric co-membership matrices.
}
\description{
Runs Partitioning Around Medoids (PAM) clustering using implementation from
\code{\link[cluster]{pam}}. This is also known as the k-medoids algorithm.
This function is not using stability.
}
\details{
Faster implementations of the algorithm can be chosen via the
  parameter \code{pamonce} (see \code{\link[cluster]{pam}}).
}
\examples{

# Data simulation
set.seed(1)
simul <- SimulateClustering(n = c(10, 10), pk = 50)

# PAM clustering
mypam <- PAMClustering(xdata = simul$data, nc = 1:20)
mypam$comembership[, , 5]

# Using another distance metric: argument passed to stats::dist()
mypam <- PAMClustering(xdata = simul$data, nc = 1:20, method = "manhattan")
mypam$comembership[, , 5]

# Using specific medoids: arguments passed to cluster::pam()
mypam <- PAMClustering(
  xdata = simul$data, nc = 5,
  medoids = 1:5, do.swap = FALSE
)
plot(Graph(
  adjacency = mypam$comembership[, , 1],
  node_colour = c(rep("red", 5), rep("blue", 15)),
  satellites = TRUE
))
}
\seealso{
Other clustering algorithms: 
\code{\link{CLARAClustering}()},
\code{\link{DBSCANClustering}()},
\code{\link{GMMClustering}()},
\code{\link{HierarchicalClustering}()},
\code{\link{KMeansClustering}()}
}
\concept{clustering algorithms}
