% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_models.R
\name{PenalisedRegression}
\alias{PenalisedRegression}
\title{Penalised regression}
\usage{
PenalisedRegression(xdata, ydata, Lambda = NULL, family, ...)
}
\arguments{
\item{xdata}{matrix of predictors with observations as rows and variables as
columns.}

\item{ydata}{optional vector or matrix of outcome(s).}

\item{Lambda}{matrix of parameters controlling the level of sparsity in the
underlying feature selection algorithm specified in \code{implementation}.
With \code{implementation="glmnet"}, \code{Lambda} contains penalty
parameters. If \code{Lambda=NULL}, \code{\link{LambdaGridRegression}} is
used to define a relevant grid. If \code{implementation} is not set to
\code{"glmnet"}, \code{Lambda} must be provided.}

\item{family}{type of regression model. If \code{implementation="glmnet"},
this argument is defined as in \code{\link[glmnet]{glmnet}}. Possible
values include \code{"gaussian"} (linear regression), \code{"binomial"}
(logistic regression), \code{"multinomial"} (multinomial regression), and
\code{"cox"} (survival analysis).}

\item{...}{additional parameters passed to \code{\link[glmnet]{glmnet}}.}
}
\value{
A list with: \item{selected}{matrix of binary selection status. Rows
  correspond to different model parameters. Columns correspond to
  predictors.} \item{beta_full}{array of model coefficients. Rows correspond
  to different model parameters. Columns correspond to predictors. Indices
  along the third dimension correspond to outcome variable(s).}
}
\description{
Runs penalised regression using implementation from
\code{\link[glmnet]{glmnet}}. This function is not using stability.
}
\examples{

# Data simulation
set.seed(1)
simul <- SimulateRegression(pk = 50)

# Running the LASSO
mylasso <- PenalisedRegression(
  xdata = simul$X, ydata = simul$Y,
  Lambda = c(0.1, 0.2), family = "gaussian"
)
}
\seealso{
\code{\link{SelectionAlgo}}, \code{\link{VariableSelection}}

Other underlying algorithm functions: 
\code{\link{GraphicalAlgo}()},
\code{\link{PenalisedGraphical}()},
\code{\link{SelectionAlgo}()}
}
\concept{underlying algorithm functions}
