% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_models.R
\name{PenalisedSEM}
\alias{PenalisedSEM}
\title{Penalised Structural Equation Model}
\usage{
PenalisedSEM(
  xdata,
  adjacency,
  residual_covariance = NULL,
  Lambda,
  n_convergence = 500,
  ...
)
}
\arguments{
\item{xdata}{matrix with observations as rows and variables as columns.
Column names must be defined and in line with the row and column names of
\code{adjacency}.}

\item{adjacency}{binary adjacency matrix of the Directed Acyclic Graph
(transpose of the asymmetric matrix A in Reticular Action Model notation).
The row and column names of this matrix must be defined.}

\item{residual_covariance}{binary and symmetric matrix encoding the nonzero
entries in the residual covariance matrix (symmetric matrix S in Reticular
Action Model notation). By default, this is the identity matrix (no
residual covariance).}

\item{Lambda}{matrix of parameters controlling the level of sparsity.}

\item{n_convergence}{maximum number of attempts to convergence.}

\item{...}{additional parameters passed to \code{\link[regsem]{regsem}}.}
}
\value{
A list with: \item{selected}{matrix of binary selection status. Rows
  correspond to different regularisation parameters. Columns correspond to
  different parameters to estimated.} \item{beta_full}{array of model
  coefficients. Rows correspond to different regularisation parameters.
  Columns correspond to different parameters to estimated.}
}
\description{
Runs penalised Structural Equation Modelling using implementation from
\code{\link[regsem]{regsem}}. This function is not using stability.
}
\examples{
\donttest{
# Definition of simulated effects
dag <- LayeredDAG(layers = c(3, 2, 3))
theta <- dag
theta[2, 4] <- 0

# Data simulation
set.seed(1)
simul <- SimulateStructural(theta = theta)

# Running regularised SEM
mysem <- PenalisedSEM(
  xdata = simul$data, adjacency = dag,
  Lambda = LambdaSequence(lmax = 1, lmin = 0.01, cardinal = 5)
)
mysem$selected

# Obtaining results in matrix format
LavaanMatrix(vect = mysem$selected[3, ], adjacency = dag)

# Data simulation with latent variables
set.seed(1)
pk <- c(3, 2, 3)
simul <- SimulateStructural(
  n = 500,
  pk = pk,
  nu_between = 0.5,
  v_sign = 1,
  n_manifest = 3,
  ev_manifest = 0.95
)

# Running regularised SEM
dag <- LayeredDAG(layers = pk, n_manifest = 3)
mysem <- PenalisedSEM(
  xdata = simul$data, adjacency = dag,
  Lambda = LambdaSequence(lmax = 1, lmin = 0.01, cardinal = 5)
)
mysem$selected
}
}
\references{
\insertRef{RegSEM}{sharp}

\insertRef{lavaanBook}{sharp}
}
\seealso{
\code{\link{SelectionAlgo}}, \code{\link{VariableSelection}},
  \code{\link{LavaanMatrix}}

Other underlying algorithm functions: 
\code{\link{ClusteringAlgo}()},
\code{\link{PenalisedGraphical}()},
\code{\link{PenalisedOpenMx}()},
\code{\link{PenalisedRegression}()}
}
\concept{underlying algorithm functions}
