% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_plot.R
\name{plot.variable_selection}
\alias{plot.variable_selection}
\title{Plot of selection proportions}
\usage{
\method{plot}{variable_selection}(
  x,
  col = c("red", "grey"),
  col.axis = NULL,
  col.thr = "darkred",
  lty.thr = 2,
  n_predictors = NULL,
  ...
)
}
\arguments{
\item{x}{output of \code{\link{VariableSelection}}.}

\item{col}{vector of colours by stable selection status.}

\item{col.axis}{optional vector of label colours by stable selection status.}

\item{col.thr}{threshold colour.}

\item{lty.thr}{threshold line type as \code{lty} in
\code{\link[graphics]{par}}.}

\item{n_predictors}{number of predictors to display.}

\item{...}{additional plotting arguments (see \code{\link[graphics]{par}}).}
}
\value{
A plot.
}
\description{
Makes a barplot of selection proportions in decreasing order. See examples in
\code{\link{VariableSelection}}.
}
\seealso{
\code{\link{VariableSelection}}
}
