% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{PlotROC}
\alias{PlotROC}
\title{Receiver Operating Characteristic (ROC) curve}
\usage{
PlotROC(
  roc,
  col = "red",
  col_band = "red",
  alpha = 0.5,
  lwd = 1,
  lty = 1,
  quantiles = c(0.05, 0.95),
  add = FALSE
)
}
\arguments{
\item{roc}{output from \code{\link{ROC}} or
\code{\link{ExplanatoryPerformance}}.}

\item{col}{colour of the point-wise average curve.}

\item{col_band}{colour of the band defined by point-wise \code{quantiles}.}

\item{alpha}{level of opacity for the band.}

\item{lwd}{line width for the point-wise average curve, as in
\code{\link{par}}.}

\item{lty}{line type for the point-wise average curve, as in
\code{\link{par}}.}

\item{quantiles}{point-wise quantiles of the performances defining the band.}

\item{add}{logical indicating if the curve should be added to the current
plot.}
}
\value{
A plot.
}
\description{
Plots the True Positive Rate (TPR) as a function of the False Positive Rate
(FPR) for different thresholds in predicted probabilities. If the results
from multiple ROC analyses are provided (e.g. output from
\code{\link{ExplanatoryPerformance}} with large \code{K}), the point-wise
average is represented and flanked by a transparent band defined by
point-wise \code{quantiles}.
}
\examples{
\dontrun{
# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 500, pk = 10, family = "binomial")

# Balanced split: 50\% variable selection set and 50\% for evaluation of performances
ids_train <- Resample(data = simul$ydata, tau = 0.5)
xtrain <- simul$xdata[ids_train, ]
ytrain <- simul$ydata[ids_train, ]
xtest <- simul$xdata[-ids_train, ]
ytest <- simul$ydata[-ids_train, ]

# Stability selection
stab <- VariableSelection(xdata = xtrain, ydata = ytrain, family = "binomial")

# Evaluation of the performances on recalibrated models (K=1)
roc <- ExplanatoryPerformance(
  xdata = xtest, ydata = ytest,
  stability = stab, n_thr = NULL
)
PlotROC(roc)

# Using more recalibration/test splits
roc <- ExplanatoryPerformance(
  xdata = xtest, ydata = ytest,
  stability = stab, K = 100
)
PlotROC(roc)

# Comparison with saturated model
roc <- ExplanatoryPerformance(
  xdata = xtest, ydata = ytest,
  family = "binomial", K = 100
)
PlotROC(roc, col = "blue", col_band = "blue", add = TRUE)
}
}
\seealso{
\code{\link{VariableSelection}}, \code{\link{Recalibrate}}

Other prediction performance functions: 
\code{\link{ExplanatoryPerformance}()},
\code{\link{ROC}()}
}
\concept{prediction performance functions}
