% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{PredictPLS}
\alias{PredictPLS}
\title{Partial Least Squares predictions}
\usage{
PredictPLS(xdata, pls)
}
\arguments{
\item{xdata}{matrix of predictors with observations as rows and variables as
columns.}

\item{pls}{output of \code{\link{PLS}}.}
}
\value{
A list with: \item{Wmat}{matrix of X-weights.} \item{Wstar}{matrix of
  transformed X-weights.} \item{Pmat}{matrix of X-loadings.}
  \item{Tmat}{matrix of X-scores.} \item{Umat}{matrix of Y-scores.}
  \item{Qmat}{matrix needed for predictions.} \item{Rmat}{matrix needed for
  predictions.} \item{meansX}{vector used for centering of predictors, needed
  for predictions.} \item{sigmaX}{vector used for scaling of predictors,
  needed for predictions.} \item{meansY}{vector used for centering of
  outcomes, needed for predictions.} \item{sigmaY}{vector used for scaling of
  outcomes, needed for predictions.} \item{methods}{a list with \code{family}
  and \code{scale} values used for the run.} \item{params}{a list with
  \code{selectedX} and \code{selectedY} values used for the run.}
}
\description{
Computes predicted values from a Partial Least Squares (PLS) model in
regression mode applied on \code{xdata}. This function is using the algorithm
implemented in \code{\link[mixOmics]{predict.pls}}.
}
\examples{
# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 100, pk = c(5, 5, 5), family = "gaussian")
x <- simul$xdata
y <- simul$ydata

# PLS
mypls <- PLS(xdata = x, ydata = y, ncomp = 3)

# Predicted values
predicted <- PredictPLS(xdata = x, pls = mypls)
}
\seealso{
\code{\link{PLS}}
}
