% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranM.R
\name{ranM}
\alias{ranM}
\title{Prepare a random matrix}
\usage{
ranM(scdata, p)
}
\arguments{
\item{scdata}{input single-cell expression matrix}

\item{p}{the dimension to be reduced to}
}
\description{
This function is to prepare a random matrix to be used in parallel computation, so that the same random matrix will be applied to different partitions/groups of datasets in the same application of random projection.
}
