% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranM2.R
\name{ranM2}
\alias{ranM2}
\title{Prepare a random matrix}
\usage{
ranM2(m, p, seedn)
}
\arguments{
\item{m}{number of genes for the input scRNA-seq matrix}

\item{p}{the dimension to be reduced to}
}
\description{
This function is to prepare a random matrix to be used in parallel computation, so that the same random matrix will be applied to different partitions/groups of datasets in the same application of random projection. The difference between ranM and ranM2 is that the former requires the input expression matrix, whereas the latter only requires the number of genes.
}
