% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanatory_performance.R
\name{Rates}
\alias{Rates}
\title{True and False Positive Rates}
\usage{
Rates(observed, predicted, thr)
}
\arguments{
\item{observed}{observed binary status.}

\item{predicted}{predicted score.}

\item{thr}{threshold for predicted probabilities.}
}
\value{
True and False Positive Rates (TPR and FPR, respectively).
}
\description{
Computes the True and False Positive Rates by comparing the true (observed) and
predicted status. The predicted status is obtained by applying a threshold on
the predicted scores.
}
\keyword{internal}
