% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{ROC}
\alias{ROC}
\title{Receiver Operating Characteristic (ROC)}
\usage{
ROC(predicted, observed, n_thr = NULL)
}
\arguments{
\item{predicted}{numeric predicted scores.}

\item{observed}{factor encoding the observed binary status.}

\item{n_thr}{number of thresholds to use to construct the ROC curve. For
faster computations on large data, values below \code{length(x)-1} can be
used.}
}
\value{
A list with: \item{TPR}{True Positive Rate.} \item{FPR}{False
  Positive Rate.} \item{AUC}{Area Under the Curve.}
}
\description{
Computes the True and False Positive Rates (TPR and FPR, respectively) and
Area Under the Curve (AUC) by comparing the true (observed) and predicted
status using a range of thresholds on the predicted score.
}
\examples{
\dontrun{

# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 500, pk = 20, family = "binomial")

# Balanced training/test split
ids_train <- Resample(data = simul$ydata, tau = 0.5)
xtrain <- simul$xdata[ids_train, , drop = FALSE]
ytrain <- simul$ydata[ids_train, , drop = FALSE]
ids_recalib <- Resample(data = simul$ydata[-ids_train, , drop = FALSE], tau = 0.5)
xrecalib <- simul$xdata[ids_recalib, , drop = FALSE]
yrecalib <- simul$ydata[ids_recalib, , drop = FALSE]
xtest <- simul$xdata[-ids_recalib, ]
ytest <- simul$ydata[-ids_recalib, ]

# Stability selection and recalibration
stab <- VariableSelection(xdata = xrecalib, ydata = yrecalib, family = "binomial")
recalibrated <- Recalibrate(xdata = xtrain, ydata = ytrain, stability = stab)

# ROC analysis
predicted <- predict(recalibrated, newdata = as.data.frame(xtest))
roc <- ROC(predicted = predicted, observed = ytest)
PlotROC(roc)
}

}
\seealso{
Other prediction performance functions: 
\code{\link{ExplanatoryPerformance}()},
\code{\link{Incremental}()},
\code{\link{PlotIncremental}()},
\code{\link{PlotROC}()}
}
\concept{prediction performance functions}
