% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPmat.R
\name{RPmat}
\alias{RPmat}
\title{Obtain a random matrix}
\usage{
RPmat(scdata, p)
}
\arguments{
\item{scdata}{input single-cell expression matrix}

\item{p}{the dimension to be reduced to}
}
\description{
This function is to generate a random matrix according to a sparse-version Achlioptas distribution (see our paper for details), given the input matrix and the dimension we want to reduce to.
}
\examples{
newE = RPmat(E, p)

}
