% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{SelectedVariables}
\alias{SelectedVariables}
\title{Set of stably selected variables}
\usage{
SelectedVariables(stability, argmax_id = NULL)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}}.}

\item{argmax_id}{matrix of parameter IDs.
If argmax_id=NULL, the calibrated set is returned.}
}
\value{
a binary vector encoding the selection status
of the variables.
}
\description{
Builds the (calibrated) set of stably selected variables.
}
\examples{
# Data simulation
set.seed(1)
simul=SimulateRegression(pk=50)

# Stability selection
stab=VariableSelection(xdata=simul$X, ydata=simul$Y)

# Calibrated set
A=SelectedVariables(stab)

# User-defined parameters
myids=matrix(c(80,10), nrow=1)
stab$Lambda[myids[1],1] # corresponding penalty
stab$params$pi_list[myids[2]] # corresponding threshold
A=SelectedVariables(stab, argmax_id=myids)

}
