% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{SelectionAlgo}
\alias{SelectionAlgo}
\title{Variable selection algorithm}
\usage{
SelectionAlgo(xdata, ydata, Lambda, family, implementation = "glmnet", ...)
}
\arguments{
\item{xdata}{matrix of predictors with observations as rows and variables as
columns.}

\item{ydata}{vector or matrix of outcome(s).}

\item{Lambda}{matrix of parameters controlling the level of sparsity in the
underlying feature selection algorithm specified in \code{implementation}.
With \code{implementation="glmnet"}, \code{Lambda} contains penalty
parameters. If \code{Lambda=NULL}, \code{\link{LambdaGridRegression}} is
used to define a relevant grid. If \code{implementation} is not set to
\code{"glmnet"}, \code{Lambda} must be provided.}

\item{family}{type of regression model. If \code{implementation="glmnet"},
this argument is defined as in \code{\link[glmnet]{glmnet}}. Possible
values include \code{"gaussian"} (linear regression), \code{"binomial"}
(logistic regression), \code{"multinomial"} (multinomial regression), and
\code{"cox"} (survival analysis).}

\item{implementation}{character string indicating the name of the function to
use for variable selection. If \code{implementation="glmnet"},
\code{\link[glmnet]{glmnet}} is used for regularised regression.
Alternatively, a function with arguments \code{xdata}, \code{ydata},
\code{Lambda}, \code{family} and \code{...}, and returning a list of two
matrices named \code{selected} and \code{beta_full} of the correct
dimensions can be used (more details in \code{\link{SelectionAlgo}}).}

\item{...}{additional parameters passed to the function provided in
\code{implementation}.}
}
\value{
A list with: \item{selected}{matrix of binary selection status. Rows
  correspond to different model parameters. Columns correspond to
  predictors.} \item{beta_full}{array of model coefficients. Rows correspond
  to different model parameters. Columns correspond to predictors. Indices
  along the third dimension correspond to outcome variable(s).}
}
\description{
Runs the variable selection algorithm specified in the argument
\code{implementation} and returns matrices of model coefficients. This
function is not using stability.
}
\examples{
\dontrun{

# Data simulation
set.seed(1)
simul <- SimulateRegression(pk = 50)

# Running the LASSO
mylasso <- SelectionAlgo(
  xdata = simul$X, ydata = simul$Y,
  Lambda = c(0.1, 0.2), family = "gaussian"
)

# Simulation of additional outcomes
set.seed(2)
Y <- cbind(simul$Y, matrix(rnorm(nrow(simul$Y) * 2), ncol = 2))

# Running multivariate Gaussian LASSO
mylasso <- SelectionAlgo(
  xdata = simul$X, ydata = Y,
  Lambda = c(0.1, 0.2), family = "mgaussian"
)
str(mylasso)
}

}
\seealso{
\code{\link{VariableSelection}}

Other underlying algorithm functions: 
\code{\link{GraphicalAlgo}()}
}
\concept{underlying algorithm functions}
