% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{SelectionProportionsGraphical}
\alias{SelectionProportionsGraphical}
\title{Selection proportions (graphical model)}
\usage{
SelectionProportionsGraphical(stability, argmax_id = NULL)
}
\arguments{
\item{stability}{output of \code{\link{GraphicalModel}}.}

\item{argmax_id}{optional matrix of parameter IDs. If \code{argmax_id=NULL},
the calibrated adjacency matrix is returned.}
}
\value{
A symmetric matrix.
}
\description{
Extracts the selection proportions of the (calibrated) stability selection
model.
}
\keyword{internal}
