% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{SelectionProportionsRegression}
\alias{SelectionProportionsRegression}
\title{Selection proportions (variable selection)}
\usage{
SelectionProportionsRegression(stability, argmax_id = NULL)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}}.}

\item{argmax_id}{matrix of parameter IDs.
If argmax_id=NULL, selection proportions of the calibrated
stability selection model are returned.}
}
\value{
a vector (variable selection)
of selection proportions.
}
\description{
Extracts the selection proportions of the
(calibrated) stability selection model.
}
\keyword{internal}
