% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHARP.R
\name{SHARP_small}
\alias{SHARP_small}
\title{Run SHARP for single-cell RNA data clustering}
\usage{
SHARP_small(E, ncells, K, p)
}
\arguments{
\item{E}{input single-cell expression matrix}

\item{K}{number of applications of random projection}

\item{p}{the dimension to be reduced to}

\item{gtc}{the ground-truth clusters}
}
\description{
SHARP: \strong{S}ingle-cell RNA-Seq \strong{H}yper-fast and \strong{A}ccurate clustering via ensemble \strong{R}andom \strong{P}rojection.
}
\examples{
enresults = SHARP(scExp)

}
