% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMetaC.R
\name{sMetaC}
\alias{sMetaC}
\title{similarity-based ensemble meta-clustering}
\usage{
sMetaC(rerowColor, sE1, folds)
}
\arguments{
\item{nC}{a m*n matrix, where m is the number of cells, n is the number of clustering algorithms, and the (i,j)-element of nC represents the cluter for the i-th cell by the j-th clutering predictor.}
}
\description{
This function is to do similarity-based ensemble clustering for combining the results from different partitions of single cells. The similarity is measured by the group-group correlations.
}
\examples{
finalrowColor = sMetaC(rerowColor, sE1, folds)

}
