% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{SparseGroupPLS}
\alias{SparseGroupPLS}
\title{Sparse group Partial Least Squares}
\usage{
SparseGroupPLS(
  x,
  y,
  group_x,
  group_y = NULL,
  lambda,
  family = "gaussian",
  ncomp = 1,
  keepX_previous = NULL,
  keepY = NULL,
  ...
)
}
\arguments{
\item{x}{matrix of predictors with observations as rows and variables as
columns.}

\item{y}{vector or matrix of outcome(s).}

\item{group_x}{vector encoding the grouping structure among predictors. This
argument indicates the number of variables in each group.}

\item{group_y}{optional vector encoding the grouping structure among
outcomes. This argument indicates the number of variables in each group.}

\item{lambda}{matrix of parameters controlling the number of selected groups
at current component, as defined by \code{ncomp}.}

\item{family}{type of PLS model. If \code{family="gaussian"}, a sparse PLS
model as defined in \code{\link[sgPLS]{sPLS}} is run (for continuous
outcomes). If \code{family="binomial"}, a PLS-DA model as defined in
\code{\link[sgPLS]{sPLSda}} is run (for categorical outcomes).}

\item{ncomp}{number of components.}

\item{keepX_previous}{number of selected groups in previous components. Only
used if \code{ncomp > 1}. The argument \code{keepX} in
\code{\link[sgPLS]{sgPLS}} is obtained by concatenating
\code{keepX_previous} and \code{lambda}.}

\item{keepY}{number of selected groups of outcome variables. This argument is
defined as in \code{\link[sgPLS]{sgPLS}}. Only used if
\code{family="gaussian"}.}

\item{...}{additional arguments to be passed to \code{\link[sgPLS]{sgPLS}} or
\code{\link[sgPLS]{sgPLSda}}.}
}
\value{
A list with: \item{selected}{matrix of binary selection status. Rows
  correspond to different model parameters. Columns correspond to
  predictors.} \item{beta_full}{array of model coefficients. Rows correspond
  to different model parameters. Columns correspond to predictors (starting
  with "X") or outcomes (starting with "Y") variables for different
  components (denoted by "PC").}
}
\description{
Runs a sparse group Partial Least Squares model using implementation from
\code{\link[sgPLS]{sgPLS-package}}. This function is not using stability.
}
\examples{
## Sparse group PLS
# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 100, pk = 50, family = "gaussian")
ydata <- cbind(simul$Y, matrix(rnorm(100 * 3), ncol = 3))
colnames(ydata) <- paste0("outcome", 1:4)
x <- simul$X
y <- ydata

# Running sgPLS with 1 group and sparsity of 0.5
mypls <- SparseGroupPLS(
  x = x, y = y, lambda = 1, family = "gaussian",
  group_x = c(20, 15, 25), alpha.x = 0.5
)

# Running sgPLS with groups on outcomes
mypls <- SparseGroupPLS(
  x = x, y = y, lambda = 1, family = "gaussian",
  group_x = c(20, 15, 25), alpha.x = 0.5,
  group_y = c(2, 2), keepY = 1, alpha.y = 0.9
)

## Sparse group PLS-DA
# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 100, pk = 50, family = "binomial")

# Running sgPLS-DA with 1 group and sparsity of 0.9
mypls <- SparseGroupPLS(
  x = simul$X, y = simul$Y, lambda = 1, family = "binomial",
  group_x = c(20, 15, 25), alpha.x = 0.9
)
}
\seealso{
\code{\link{SelectionAlgo}}, \code{\link{VariableSelection}}

Other PLS functions: 
\code{\link{SparsePLS}()}
}
\concept{PLS functions}
