% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{SparsePCA}
\alias{SparsePCA}
\title{Sparse Principal Component Analysis}
\usage{
SparsePCA(xdata, Lambda, ncomp = 1, keepX_previous = NULL, ...)
}
\arguments{
\item{xdata}{data matrix with observations as rows and variables as columns.}

\item{Lambda}{matrix of parameters controlling the number of selected
variables at current component, as defined by \code{ncomp}.}

\item{ncomp}{number of components.}

\item{keepX_previous}{number of selected predictors in previous components.
Only used if \code{ncomp > 1}. The argument \code{keepX} in
\code{\link[mixOmics]{spca}} is obtained by concatenating
\code{keepX_previous} and \code{Lambda}.}

\item{...}{additional arguments to be passed to \code{\link[mixOmics]{spca}}.}
}
\value{
A list with: \item{selected}{matrix of binary selection status. Rows
  correspond to different model parameters. Columns correspond to
  predictors.} \item{beta_full}{array of model coefficients. Rows correspond
  to different model parameters. Columns correspond to predictors (starting
  with "X") or outcomes (starting with "Y") variables for different
  components (denoted by "PC").}
}
\description{
Runs a sparse Principal Component Analysis model using implementation from
\code{\link[elasticnet]{spca}}. This function is not using stability.
}
\examples{
# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 100, pk = 50, family = "gaussian")
x <- simul$xdata

# Sparse PCA
mypca <- SparsePCA(xdata = x, ncomp = 2, Lambda = c(1, 2), keepX_previous = 10)
}
\seealso{
\code{\link{SelectionAlgo}}, \code{\link{VariableSelection}}

Other penalised dimensionality reduction functions: 
\code{\link{GroupPLS}()},
\code{\link{SparseGroupPLS}()},
\code{\link{SparsePLS}()}
}
\concept{penalised dimensionality reduction functions}
