% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{SparsePLS}
\alias{SparsePLS}
\title{Sparse Partial Least Squares}
\usage{
SparsePLS(
  xdata,
  ydata,
  Lambda,
  family = "gaussian",
  ncomp = 1,
  keepX_previous = NULL,
  keepY = NULL,
  ...
)
}
\arguments{
\item{xdata}{matrix of predictors with observations as rows and variables as
columns.}

\item{ydata}{vector or matrix of outcome(s).}

\item{Lambda}{matrix of parameters controlling the number of selected
predictors at current component, as defined by \code{ncomp}.}

\item{family}{type of PLS model. If \code{family="gaussian"}, a sparse PLS
model as defined in \code{\link[sgPLS]{sPLS}} is run (for continuous
outcomes). If \code{family="binomial"}, a PLS-DA model as defined in
\code{\link[sgPLS]{sPLSda}} is run (for categorical outcomes).}

\item{ncomp}{number of components.}

\item{keepX_previous}{number of selected predictors in previous components.
Only used if \code{ncomp > 1}. The argument \code{keepX} in
\code{\link[sgPLS]{sPLS}} is obtained by concatenating
\code{keepX_previous} and \code{Lambda}.}

\item{keepY}{number of selected outcome variables. This argument is defined
as in \code{\link[sgPLS]{sPLS}}. Only used if \code{family="gaussian"}.}

\item{...}{additional arguments to be passed to \code{\link[sgPLS]{sPLS}} or
\code{\link[sgPLS]{sPLSda}}.}
}
\value{
A list with: \item{selected}{matrix of binary selection status. Rows
  correspond to different model parameters. Columns correspond to
  predictors.} \item{beta_full}{array of model coefficients. Rows correspond
  to different model parameters. Columns correspond to predictors (starting
  with "X") or outcomes (starting with "Y") variables for different
  components (denoted by "PC").}
}
\description{
Runs a sparse Partial Least Squares model using implementation from
\code{\link[sgPLS]{sgPLS-package}}. This function is not using stability.
}
\examples{
## Sparse PLS
# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 100, pk = 50, family = "gaussian")
ydata <- cbind(simul$Y, matrix(rnorm(100 * 3), ncol = 3))
colnames(ydata) <- paste0("outcome", 1:4)
x <- simul$X
y <- ydata

# Running sPLS with 2 X-variables and 1 Y-variable
mypls <- SparsePLS(xdata = x, ydata = y, Lambda = 2, family = "gaussian", keepY = 1)

## Sparse PLS-DA
# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 200, pk = 20, family = "binomial")

# Running sPLS-DA with 2 X-variables and 1 Y-variable
mypls <- SparsePLS(xdata = simul$X, ydata = simul$Y, Lambda = 2, family = "binomial")
}
\seealso{
\code{\link{SelectionAlgo}}, \code{\link{VariableSelection}}

Other PLS functions: 
\code{\link{GroupPLS}()},
\code{\link{SparseGroupPLS}()}
}
\concept{PLS functions}
