% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHARP.R
\name{testlog}
\alias{testlog}
\title{Test if log-transformation is necessary.}
\usage{
testlog(scExp, ncells, p, sncells, n.cores)
}
\arguments{
\item{scExp}{input single-cell expression matrix}

\item{ncells}{number of cells in the expression matrix}

\item{p}{number of reduced dimensions by random projection}

\item{sncells}{number of cells randomly selected for checking log-transform. By default, sncells = 100}
}
\description{
Somtimes log-transformation can boost the clustering performance; however, it is not always the case. By this function, SHARP can automatically determine whether a log-transformation is necessary.
}
\examples{
flag = testlog(scExp)

}
