% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{TuneExplainedVarianceReg}
\alias{TuneExplainedVarianceReg}
\title{Tuning function (regression)}
\usage{
TuneExplainedVarianceReg(ev_xz, omega, tol = 0.1, q, p)
}
\arguments{
\item{ev_xz}{proportion of explained variance.}

\item{omega}{precision matrix.}

\item{tol}{requested smallest eigenvalue after transformation of the input
precision matrix.}

\item{q}{number of outcome variables.}

\item{p}{number of predictor variables.}
}
\value{
The absolute difference between the smallest eigenvalue of the
  transformed precision matrix and requested value \code{tol}.
}
\description{
Computes the absolute difference between the smallest eigenvalue and
requested one (parameter \code{tol}) for a precision matrix with predictors
and outcomes.
}
\keyword{internal}
