% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_SHARP.R
\name{visualization_SHARP}
\alias{visualization_SHARP}
\title{SHARP visualization}
\usage{
visualization_SHARP(y, label, w, filename, filetype, width = 900,
  height = 900)
}
\arguments{
\item{y}{the results after running the SHARP function}

\item{label}{the reference (or pre-defined) clustering label for the data; if not given, the resulting plot will be drawn in black/white color.}

\item{w}{the relative weight of clustering-result based matrix over ensemble random-projection based matrix; by default, w = 2.}

\item{filename}{the name of the file to which the output figure will be saved; if not given, the output figure will be saved in "vi_SHARP.pdf"}

\item{filetype}{the type of the output file. Suggested file types are PDF or PNG, while other common types (e.g., JPEG or TIFF) are also acceptable. If not given, the type will be determined as follows: when the number of single cells is less than 5000, the type will be PDF; otherwise, it will be PNG.}

\item{width, }{height the width and height of the outpuf figure.}
}
\description{
This function is to visualize the data in 2D scatter plots similar to tSNE. The difference between SHARP and tSNE in terms of visualization is that the latter uses the original high-dimensional data, whereas the former uses the clustered output to form a small-dimensional feature matrix for visualization.
}
\examples{
y = SHARP(scExp)
visualization_SHARP(y)

}
