% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{WeightBoxplot}
\alias{WeightBoxplot}
\title{Stable attribute weights}
\usage{
WeightBoxplot(
  stability,
  at = NULL,
  argmax_id = NULL,
  col = NULL,
  boxwex = 0.3,
  xlab = "",
  ylab = "Weight",
  cex.lab = 1.5,
  las = 3,
  frame = "F",
  add = FALSE,
  ...
)
}
\arguments{
\item{stability}{output of \code{\link{Clustering}}.}

\item{at}{coordinates along the x-axis (more details in
\code{\link[graphics]{boxplot}}).}

\item{argmax_id}{optional matrix of parameter IDs. If \code{argmax_id=NULL},
the calibrated model is used.}

\item{col}{optional vector of colours.}

\item{boxwex}{box width (more details in \code{\link[graphics]{boxplot}}).}

\item{xlab}{label of the x-axis.}

\item{ylab}{label of the y-axis.}

\item{cex.lab}{font size for labels.}

\item{las}{orientation of labels on the x-axis (see
\code{\link[graphics]{par}}).}

\item{frame}{logical indicating if the box around the plot should be drawn
(more details in \code{\link[graphics]{boxplot}}).}

\item{add}{logical indicating if the boxplot should be added to the current
plot.}

\item{...}{additional parameters passed to \code{\link[graphics]{boxplot}}).}
}
\value{
A boxplot.
}
\description{
Creates a boxplots of the distribution of (calibrated) median attribute
weights obtained from the COSA algorithm across the subsampling iterations.
}
\examples{
\donttest{
# Data simulation
set.seed(2)
simul <- SimulateClustering(
  n = c(30, 30, 30), pk = 15,
  theta_xc = c(rep(1, 5), rep(0, 10)),
  ev_xc = c(rep(0.95, 5), rep(0, 10)),
)
plot(simul)

# Consensus weighted clustering
stab <- Clustering(
  xdata = simul$data, Lambda = 0.5
)
WeightBoxplot(stability = stab)
}

}
\seealso{
\code{\link{Clustering}}

Other calibration functions: 
\code{\link{Adjacency}()},
\code{\link{ArgmaxId}()},
\code{\link{Argmax}()},
\code{\link{CalibrationPlot}()},
\code{\link{Clusters}()},
\code{\link{ConsensusMatrix}()},
\code{\link{SelectedVariables}()},
\code{\link{SelectionProportions}()}
}
\concept{calibration functions}
