% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMetaC.R
\name{wMetaC}
\alias{wMetaC}
\title{weighted ensemble clustering}
\usage{
wMetaC(nC, enN.cluster)
}
\arguments{
\item{nC}{a m*n matrix, where m is the number of cells, n is the number of clustering algorithms, and the (i,j)-element of nC represents the cluter for the i-th cell by the j-th clutering predictor.}
}
\description{
This function is to do weighted ensemble clustering for meta-clustering.
}
\examples{
finalrowColor = wMetaC(enrp)

}
