% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{Argmax}
\alias{Argmax}
\title{Calibrated parameters}
\usage{
Argmax(stability)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}} or
\code{\link{GraphicalModel}}.}
}
\value{
A matrix of parameter values. In multi-block graphical modelling,
  rows correspond to different blocks.
}
\description{
Extracts calibrated parameter values in stability selection.
}
\examples{
\dontrun{

# Data simulation
set.seed(1)
simul <- SimulateGraphical(pk = 20)

# Stability selection
stab <- GraphicalModel(xdata = simul$data)

# Extracting calibrated parameters
args <- Argmax(stab)
}

}
\seealso{
\code{\link{VariableSelection}}, \code{\link{GraphicalModel}}

Other calibration functions: 
\code{\link{Adjacency}()},
\code{\link{ArgmaxId}()},
\code{\link{CalibrationPlot}()},
\code{\link{Clusters}()},
\code{\link{SelectedVariables}()},
\code{\link{SelectionProportions}()}
}
\concept{calibration functions}
