% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{Argmax}
\alias{Argmax}
\title{Calibrated parameters}
\usage{
Argmax(stability)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}} or
\code{\link{GraphicalModel}}.}
}
\value{
a matrix of (block-specific) parameters. In multi-block graphical
  modelling, rows correspond to different blocks.
}
\description{
Extracts calibrated parameters in stability selection.
}
\examples{
# Data simulation
set.seed(1)
simul <- SimulateGraphical(pk = 20)

# Stability selection
stab <- GraphicalModel(data = simul$data)

# Extracting calibrated parameters
args <- Argmax(stab)

}
\seealso{
Other calibration functions: 
\code{\link{Adjacency}()},
\code{\link{ArgmaxId}()},
\code{\link{CalibrationPlot}()},
\code{\link{SelectedVariables}()},
\code{\link{SelectionProportions}()}
}
\concept{calibration functions}
