% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{ArgmaxId}
\alias{ArgmaxId}
\title{Calibrated parameter IDs}
\usage{
ArgmaxId(stability = NULL, S = NULL)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}} or
\code{\link{GraphicalModel}}. If stability=NULL, argument "S" must be
provided.}

\item{S}{matrix of stability scores obtained with different combinations of
parameters. Rows correspond to different penalty parameters and columns
correspond to different thresholds in selection proportions. If S=NULL,
argument "stability" must be provided.}
}
\value{
a matrix of (block-specific) parameter IDs. In multi-block graphical
  modelling, rows correspond to different blocks.
}
\description{
Extracts the IDs of calibrated parameters with respect to the grids provided
in "Lambda" and "pi_list".
}
\examples{
# Data simulation
set.seed(1)
simul <- SimulateGraphical(pk = 20)

# Stability selection
stab <- GraphicalModel(data = simul$data)

# Extracting IDs of calibrated parameters
ids <- ArgmaxId(stab)
stab$Lambda[ids[1], 1]
stab$params$pi_list[ids[2]]

# Link with Argmax() function
args <- Argmax(stab)
}
\seealso{
\code{\link{Argmax}}
}
