% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{Clusters}
\alias{Clusters}
\title{Stable cluster membership}
\usage{
Clusters(adjacency = NULL, argmax_id = NULL)
}
\arguments{
\item{adjacency}{adjacency matrix or output of \code{\link{GraphicalModel}}
or \code{\link{Clustering}}.}

\item{argmax_id}{optional matrix of parameter IDs. If \code{argmax_id=NULL},
the calibrated adjacency matrix is returned.}
}
\value{
A vector encoding the cluster membership.
}
\description{
Builds the (calibrated) stable clusters as connected components of the graph
defined from stable co-membership.
}
\examples{
\dontrun{

# Data simulation
set.seed(1)
simul <- SimulateGraphical(pk = 50, n = 10)

# Consensus clustering
stab <- GraphicalModel(
  xdata = simul$data,
  Lambda = seq(2, ncol(simul$data)),
  implementation = HierarchicalClustering
)

# Stable cluster membership
groups <- Clusters(stab)

# Network representation of stable co-membership
set.seed(1)
plot(Graph(CoMembership(groups),
  satellites = TRUE,
  node_colour = distinctColorPalette(length(unique(groups)))[groups]
))
}
}
\seealso{
\code{\link{Clustering}}

Other calibration functions: 
\code{\link{Adjacency}()},
\code{\link{ArgmaxId}()},
\code{\link{Argmax}()},
\code{\link{CalibrationPlot}()},
\code{\link{SelectedVariables}()},
\code{\link{SelectionProportions}()}
}
\concept{calibration functions}
