% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{Clusters}
\alias{Clusters}
\title{Stable cluster membership}
\usage{
Clusters(stability, linkage = "complete", argmax_id = NULL)
}
\arguments{
\item{stability}{output of \code{\link{Clustering}}.}

\item{linkage}{character string indicating the type of linkage used in
hierarchical clustering to define the stable clusters. Possible values
include \code{"complete"}, \code{"single"} and \code{"average"} (see
argument \code{"method"} in \code{\link[stats]{hclust}} for a full list).}

\item{argmax_id}{optional parameter ID. If \code{argmax_id=NULL}, the
calibrated model is used.}
}
\value{
A vector of cluster memberships.
}
\description{
Generates the (calibrated) stable cluster memberships.
}
\examples{
\donttest{
# Data simulation
set.seed(1)
simul <- SimulateClustering(
  n = c(30, 30, 30),
  nu_xc = 1
)
plot(simul)

# Consensus clustering
stab <- Clustering(
  xdata = simul$data
)
Clusters(stab)
}
}
\seealso{
Other calibration functions: 
\code{\link{Adjacency}()},
\code{\link{ArgmaxId}()},
\code{\link{Argmax}()},
\code{\link{CalibrationPlot}()},
\code{\link{ConsensusMatrix}()},
\code{\link{SelectedVariables}()},
\code{\link{SelectionProportions}()},
\code{\link{WeightBoxplot}()},
\code{\link{plot.clustering}()}
}
\concept{calibration functions}
