% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelisation.R
\name{Combine}
\alias{Combine}
\title{Merging stability selection outputs}
\usage{
Combine(stability1, stability2, graph = TRUE)
}
\arguments{
\item{stability1}{output from a first run of \code{\link{VariableSelection}}
or \code{\link{GraphicalModel}}.}

\item{stability2}{output from a second run of \code{\link{VariableSelection}}
or \code{\link{GraphicalModel}}.}

\item{graph}{logical indicating if \code{\link{GraphicalModel}} (graph=TRUE)
or \code{\link{VariableSelection}} (graph=FALSE) were used to generate the
"stability1" and "stability2".}
}
\value{
a single output with the same format.
}
\description{
Merges the outputs from \code{\link{VariableSelection}} or
\code{\link{GraphicalModel}}. This can be useful for parallelisation.
}
\details{
The two runs must have been done using the same data, the same grids
of parameters, the same methods (arguments "implementation", "resampling",
"PFER_method", as well as "start" for graphical models), the same "tau" and
"pk" (graphical models only), and the same thresholds in PFER and FDP, but
with different seeds. The combined output will be the equivalent of running
the model for which the number of iterations is the sum of the number of
iterations used for "stability1" and "stability2".
}
\examples{
# Data simulation
simul <- SimulateGraphical(pk = 20)

# Two runs
stab1 <- GraphicalModel(data = simul$data, seed = 1, K = 10)
stab2 <- GraphicalModel(data = simul$data, seed = 2, K = 10)

# Merging the outputs
stab <- Combine(stability1 = stab1, stability2 = stab2, graph = TRUE)
print(stab$params$K)
}
\seealso{
\code{\link{VariableSelection}}, \code{\link{GraphicalModel}}
}
