% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelisation.R
\name{Combine}
\alias{Combine}
\title{Merging stability selection outputs}
\usage{
Combine(stability1, stability2, include_beta = TRUE)
}
\arguments{
\item{stability1}{output from a first run of \code{\link{VariableSelection}}
or \code{\link{GraphicalModel}}.}

\item{stability2}{output from a second run of \code{\link{VariableSelection}}
or \code{\link{GraphicalModel}}.}

\item{include_beta}{logical indicating if the beta coefficients of visited
models should be concatenated.}
}
\value{
A single output of the same format.
}
\description{
Merges the outputs from \code{\link{VariableSelection}} or
\code{\link{GraphicalModel}}. This function can be used for parallelisation.
}
\details{
The two runs must have been done using the same data, the same grids
  of parameters, the same methods (arguments \code{implementation},
  \code{resampling}, \code{PFER_method}, as well as \code{start} for
  graphical models), the same \code{tau} and \code{pk} (graphical models
  only), and the same thresholds in PFER and FDP, but with different seeds.
  The combined output will be the equivalent of running the model for which
  the number of iterations is the sum of the number of iterations used for
  \code{stability1} and \code{stability2}.
}
\examples{

## Variable selection
# Data simulation
set.seed(1)
simul <- SimulateRegression(n = 100, pk = 50, family = "gaussian")

# Two runs
stab1 <- VariableSelection(xdata = simul$xdata, ydata = simul$ydata, seed = 1, K = 10)
stab2 <- VariableSelection(xdata = simul$xdata, ydata = simul$ydata, seed = 2, K = 10)

# Merging the outputs
stab <- Combine(stability1 = stab1, stability2 = stab2, include_beta = FALSE)
print(stab$params$K)

## Graphical modelling
# Data simulation
simul <- SimulateGraphical(pk = 20)

# Two runs
stab1 <- GraphicalModel(xdata = simul$data, seed = 1, K = 10)
stab2 <- GraphicalModel(xdata = simul$data, seed = 2, K = 10)

# Merging the outputs
stab <- Combine(stability1 = stab1, stability2 = stab2)
print(stab$params$K)

## Clustering
# Data simulation
set.seed(1)
simul <- SimulateClustering(
  n = c(5, 5, 5), pk = 100
)

# Two runs
stab1 <- Clustering(xdata = simul$data, seed = 1, K = 10)
stab2 <- Clustering(xdata = simul$data, seed = 2, K = 10)

# Merging the outputs (TODO)
}
\seealso{
\code{\link{VariableSelection}}, \code{\link{GraphicalModel}}
}
