% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CoMembership}
\alias{CoMembership}
\title{Pairwise co-membership}
\usage{
CoMembership(groups)
}
\arguments{
\item{groups}{vector of group membership.}
}
\value{
A symmetric and binary matrix.
}
\description{
Generates a symmetric and binary matrix indicating, if two items are
co-members, i.e. belong to the same cluster.
}
\examples{
# Simulated grouping structure
mygroups <- c(rep(1, 3), rep(2, 5), rep(3, 2))

# Co-membership matrix
CoMembership(mygroups)
}
