% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{Contrast}
\alias{Contrast}
\title{Matrix contrast}
\usage{
Contrast(mat, digits = 3)
}
\arguments{
\item{mat}{input matrix.}

\item{digits}{number of digits to use.}
}
\value{
A single number, the contrast of the input matrix.
}
\description{
Computes the matrix contrast, defined as the number of unique truncated
entries with a specified number of digits.
}
