% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_models.R
\name{DBSCANClustering}
\alias{DBSCANClustering}
\title{Density-Based Spatial Clustering of Applications with Noise}
\usage{
DBSCANClustering(xdata, nc = NULL, scale = TRUE, rows = TRUE, ...)
}
\arguments{
\item{xdata}{data matrix with observations as rows and variables as columns.}

\item{nc}{matrix of parameters controlling the number of clusters in the
underlying algorithm specified in \code{implementation}. If \code{nc} is
not provided, it is set to \code{seq(1, nrow(xdata))}.}

\item{scale}{logical indicating if the data should be scaled to ensure that
all variables contribute equally to the clustering of the observations.}

\item{rows}{logical indicating if clusters of rows (\code{TRUE}) or columns (\code{FALSE})
should be inferred.}

\item{...}{additional parameters passed to \code{\link[dbscan]{dbscan}}.}
}
\value{
An array with binary and symmetric co-membership matrices.
}
\description{
Runs Density-Based Spatial Clustering of Applications with Noise (DBSCAN)
algorithm using implementation from \code{\link[dbscan]{dbscan}}. This
function is not using stability.
}
\examples{

# Data simulation
set.seed(1)
simul <- SimulateClustering(n = c(10, 10), pk = 50)

# PAM clustering
mydbscan <- DBSCANClustering(
  xdata = simul$data,
  nc = seq(6, 10, by = 0.1)
)
}
\seealso{
Other clustering algorithms: 
\code{\link{CLARAClustering}()},
\code{\link{GMMClustering}()},
\code{\link{HierarchicalClustering}()},
\code{\link{KMeansClustering}()},
\code{\link{PAMClustering}()},
\code{\link{SparseHierarchicalClustering}()}
}
\concept{clustering algorithms}
