% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DummyToCategories}
\alias{DummyToCategories}
\title{Categorical from dummy variables}
\usage{
DummyToCategories(x, verbose = FALSE)
}
\arguments{
\item{x}{matrix of dummy variables.}

\item{verbose}{logical indicating if messages should be printed.}
}
\value{
A single categorical variable (numeric).
}
\description{
Generates a single categorical variable from corresponding dummy variables.
}
\keyword{internal}
