% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFER.R
\name{FDP}
\alias{FDP}
\title{False Discovery Proportion}
\usage{
FDP(PFER, stab_iter, pi)
}
\arguments{
\item{PFER}{Per Family Error Rate.}

\item{stab_iter}{matrix or vector of selection proportions.}

\item{pi}{vector of thresholds in selection proportions.}
}
\value{
the (upper-bound) of the FDP.
}
\description{
Computes the (upper-bound) of the FDP as a ratio of the (upper-bound)
of the PFER over the number of stably selected features.
In stability selection, the FDP corresponds to the expected proportion of stably selected
features that are not relevant to the outcome
(i.e. proportion of False Positives among stably selected features).
}
\concept{stability metric functions

# Simulating set of selection proportions
selprop=round(runif(n=20), digits=2)

# Computing the FDP with a threshold of 0.8
fdp=FDP(PFER=3, stab_iter=selprop, pi=0.8)}
