% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{GraphicalAlgo}
\alias{GraphicalAlgo}
\title{Graph estimation algorithm}
\usage{
GraphicalAlgo(
  x,
  pk = NULL,
  Lambda,
  Sequential_template,
  scale = TRUE,
  implementation = "glassoFast",
  start = "cold",
  ...
)
}
\arguments{
\item{x}{matrix with observations as rows and variables as columns.}

\item{pk}{vector encoding the grouping structure.
Only used for multi-block stability selection.
For this, the variables in data have to be ordered
by group and argument "pk" has to be a vector
indicating the number of variables
in each of the groups (see example below).
If pk=NULL, single-block stability selection is performed.}

\item{Lambda}{matrix of parameters controlling the underlying
feature selection algorithm specified in "implementation".
With implementation="glassoFast", these are penalty parameters
controlling the regularised model.
Lambda must be a matrix with as many columns as there are entries in "pk".}

\item{Sequential_template}{logical matrix encoding the type of procedure
to use for data with multiple blocks in stability selection graphical modelling.
For multi-block estimation, the stability selection model is
constructed as the union of block-specific stable edges estimated
while the others are weakly penalised (TRUE only for the
block currently being calibrated and FALSE for other blocks).
Other approaches with joint calibration of the blocks are allowed
(all entries are set to TRUE).}

\item{scale}{logical indicating if the correlation (if scale=TRUE)
or covariance (if scale=FALSE) matrix should be used as input
for the graphical LASSO. If implementation is not set to "glassoFast",
this argument must be used as input of the function provided instead.}

\item{implementation}{name of the function to use for graphical modelling.
With implementation="glassoFast", the function \code{\link{glassoFast}}
is used for regularised estimation of a conditional independence graph.
Alternatively, this argument can be a character string indicating the name of a function.
The function provided must use arguments called "x", "lambda" and "scale"
and return a binary and symmetric adjacency matrix (see example below).}

\item{start}{character string indicating if the algorithm should be
initialised at the estimated (inverse) covariance with previous
penalty parameters (start="warm") or not (start="cold").
Using start="warm" can speed-up the computations.
Only used for implementation="glassoFast" (see argument "start"
in \code{\link{glassoFast}}).}

\item{...}{additional parameters passed to the function provided in
"implementation".}
}
\value{
a binary and symmetric adjacency matrix.
}
\description{
Runs the algorithm for estimation of an undirected graph with no self-loops
specified in the argument "implementation"
and returns the estimated adjacency matrix.
This function is not using stability.
}
\details{
The use of the procedure from Equation (4) or (5)
is controlled by the argument "Sequential_template".
}
\examples{
# Data simulation
set.seed(1)
simul=SimulateGraphical()

# Running graphical LASSO
myglasso=GraphicalAlgo(x=simul$data, Lambda=matrix(c(0.1, 0.2),ncol=1))

}
