% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_models.R
\name{HierarchicalClustering}
\alias{HierarchicalClustering}
\title{Hierarchical clustering}
\usage{
HierarchicalClustering(xdata, nc = NULL, scale = TRUE, rows = TRUE, ...)
}
\arguments{
\item{xdata}{data matrix with observations as rows and variables as columns.}

\item{nc}{matrix of parameters controlling the number of clusters in the
underlying algorithm specified in \code{implementation}. If \code{nc}
is not provided, it is set to \code{seq(1, nrow(xdata))}.}

\item{scale}{logical indicating if the data should be scaled to ensure that
all variables contribute equally to the clustering of the observations.}

\item{rows}{logical indicating if clusters of rows (\code{TRUE}) or columns (\code{FALSE})
should be inferred.}

\item{...}{additional parameters passed to \code{\link[stats]{dist}} or
\code{\link[stats]{hclust}}.}
}
\value{
An array with binary and symmetric co-membership matrices.
}
\description{
Runs hierarchical clustering using implementation from
\code{\link[stats]{hclust}}. This function is not using stability.
}
\examples{

# Data simulation
set.seed(1)
simul <- SimulateClustering(n = c(10, 10), pk = 50)

# Hierarchical clustering
myhclust <- HierarchicalClustering(xdata = simul$data, nc = 1:20)
}
\seealso{
Other clustering algorithms: 
\code{\link{CLARAClustering}()},
\code{\link{DBSCANClustering}()},
\code{\link{GMMClustering}()},
\code{\link{KMeansClustering}()},
\code{\link{PAMClustering}()}
}
\concept{clustering algorithms}
