% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_object.R
\name{mystar}
\alias{mystar}
\title{Star-shaped nodes}
\usage{
mystar(coords, v = NULL, params)
}
\arguments{
\item{coords}{a matrix of coordinates
(see \code{\link[igraph]{add_shape}}).}

\item{v}{a vector of node IDs
(see \code{\link[igraph]{add_shape}}).}

\item{params}{node graphical parameters
(see \code{\link[igraph]{add_shape}}).}
}
\description{
Produces star-shaped nodes in an igraph object.
}
\seealso{
\code{\link[igraph]{add_shape}}
}
\keyword{internal}
