% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_models.R
\name{OpenMxMatrix}
\alias{OpenMxMatrix}
\title{Matrix from OpenMx outputs}
\usage{
OpenMxMatrix(vect, adjacency, residual_covariance = NULL)
}
\arguments{
\item{vect}{vector of coefficients to assign to entries of the matrix.}

\item{adjacency}{binary adjacency matrix of the Directed Acyclic Graph
(transpose of the asymmetric matrix A in Reticular Action Model notation).}

\item{residual_covariance}{binary and symmetric matrix encoding the nonzero
entries in the residual covariance matrix (symmetric matrix S in Reticular
Action Model notation). By default, this is the identity matrix (no
residual covariance).}
}
\value{
An asymmetric matrix.
}
\description{
Returns a matrix from output of \code{\link[OpenMx]{mxPenaltySearch}}.
}
\examples{
# Definition of the model structure
layers <- list(
  c("var1", "var2", "var3"),
  c("var4", "var5"),
  c("var6", "var7", "var8")
)
dag <- LayeredDAG(layers)

# Definition of simulated effects
theta <- dag
theta[2, 4] <- 0

# Data simulation
set.seed(1)
simul <- SimulateStructural(theta = theta)

# Running regularised SEM
mysem <- PenalisedOpenMx(
  xdata = simul$data, adjacency = dag,
  Lambda = seq(1, 50, by = 10)
)
mysem$selected

# Obtaining results in matrix format
OpenMxMatrix(vect = mysem$selected[3, ], adjacency = dag)

}
\references{
\insertRef{lavaanBook}{sharp}
}
\seealso{
\code{\link{LavaanModel}}, \code{\link{PenalisedSEM}}
}
