% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_models.R
\name{PAMClustering}
\alias{PAMClustering}
\title{Partitioning Around Medoids}
\usage{
PAMClustering(xdata, Lambda = NULL, scale = TRUE, rows = TRUE, ...)
}
\arguments{
\item{xdata}{data matrix with observations as rows and variables as columns.}

\item{Lambda}{matrix of parameters controlling the number of clusters in the
underlying algorithm specified in \code{implementation}. If \code{Lambda}
is not provided, it is set to \code{seq(1, nrow(xdata))}.}

\item{scale}{logical indicating if the data should be scaled to ensure that
all variables contribute equally to the clustering of the observations.}

\item{rows}{logical indicating if clusters of rows (\code{TRUE}) or columns (\code{FALSE})
should be conducted.}

\item{...}{additional parameters passed to \code{\link[stats]{dist}} or
\code{\link[cluster]{pam}}.}
}
\value{
An array with binary and symmetric co-membership matrices.
}
\description{
Runs Partitioning Around Medoids (PAM) clustering using implementation from
\code{\link[cluster]{pam}}. This function is not using stability.
}
\examples{

# Data simulation
set.seed(1)
simul <- SimulateClustering(n = c(10, 10), pk = 50)

# PAM clustering
mypam <- PAMClustering(xdata = simul$data, Lambda = 1:20)
}
