% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_models.R
\name{PenalisedOpenMx}
\alias{PenalisedOpenMx}
\title{Penalised Structural Equation Model (OpenMx)}
\usage{
PenalisedOpenMx(
  xdata,
  adjacency,
  penalised = NULL,
  residual_covariance = NULL,
  Lambda,
  ...
)
}
\arguments{
\item{xdata}{matrix with observations as rows and variables as columns.
Column names must be defined and in line with the row and column names of
\code{adjacency}.}

\item{adjacency}{binary adjacency matrix of the Directed Acyclic Graph
(transpose of the asymmetric matrix A in Reticular Action Model notation).
The row and column names of this matrix must be defined.}

\item{penalised}{optional binary matrix indicating which coefficients are
regularised.}

\item{residual_covariance}{binary and symmetric matrix encoding the nonzero
entries in the residual covariance matrix (symmetric matrix S in Reticular
Action Model notation). By default, this is the identity matrix (no
residual covariance).}

\item{Lambda}{matrix of parameters controlling the level of sparsity (only
the minimum, maximum and length are used).}

\item{...}{additional parameters passed to \code{\link[regsem]{regsem}}.}
}
\value{
A list with: \item{selected}{matrix of binary selection status. Rows
  correspond to different regularisation parameters. Columns correspond to
  different parameters to estimated.} \item{beta_full}{array of model
  coefficients. Rows correspond to different regularisation parameters.
  Columns correspond to different parameters to estimated.}
}
\description{
Runs penalised Structural Equation Modelling using implementation from
\code{\link[OpenMx]{OpenMx}}. This function is not using stability.
}
\examples{
# Definition of simulated effects
dag <- LayeredDAG(layers = c(3, 2, 3))
theta <- dag
theta[2, 4] <- 0

# Data simulation
set.seed(1)
simul <- SimulateStructural(theta = theta)

# Running regularised SEM
mysem <- PenalisedOpenMx(
  xdata = simul$data, adjacency = dag,
  Lambda = seq(1, 50, by = 10)
)
mysem$selected

# Obtaining results in matrix format
OpenMxMatrix(vect = mysem$selected[3, ], adjacency = dag)
theta # simulated is the same

}
\references{
\insertRef{RegSEM}{sharp}

  \insertRef{lavaanBook}{sharp}
}
\seealso{
\code{\link{SelectionAlgo}}, \code{\link{VariableSelection}},
  \code{\link{OpenMxMatrix}}

Other underlying algorithm functions: 
\code{\link{ClusteringAlgo}()},
\code{\link{PenalisedGraphical}()},
\code{\link{PenalisedRegression}()},
\code{\link{PenalisedSEM}()}
}
\concept{underlying algorithm functions}
