% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_models.R
\name{PenalisedSEM}
\alias{PenalisedSEM}
\title{Penalised Structural Equation Model}
\usage{
PenalisedSEM(
  xdata,
  adjacency,
  residual_covariance = NULL,
  Lambda,
  n_convergence = 500,
  ...
)
}
\arguments{
\item{xdata}{matrix with observations as rows and variables as columns.}

\item{adjacency}{binary adjacency matrix of the Directed Acyclic Graph
(asymmetric matrix A in Reticular Action Model notation).}

\item{residual_covariance}{binary and symmetric matrix encoding the nonzero
entries in the residual covariance matrix (symmetric matrix S in Reticular
Action Model notation). By default, this is the identity matrix (no
residual covariance).}

\item{Lambda}{matrix of parameters controlling the level of sparsity.}

\item{n_convergence}{maximum number of attempts to convergence.}

\item{...}{additional parameters passed to \code{\link[regsem]{regsem}}.}
}
\value{
A list with: \item{selected}{matrix of binary selection status. Rows
  correspond to different model parameters. Columns correspond to different
  effects.} \item{beta_full}{array of model coefficients. Rows correspond to
  different model parameters. Columns correspond to different effects.}
}
\description{
Runs penalised Structural Equation Modelling using implementation from
\code{\link[regsem]{regsem}}. This function is not using stability.
}
\examples{
# Definition of the model structure
layers <- list(
  c("var1", "var2", "var3"),
  c("var4", "var5"),
  c("var6", "var7", "var8")
)
dag <- LayeredDAG(layers)

# Definition of simulated effects
theta <- dag
theta[2, 4] <- 0

# Data simulation
set.seed(1)
simul <- SimulateSCM(theta = theta)

# Running regularised SEM
mysem <- PenalisedSEM(
  xdata = simul$data, adjacency = dag,
  Lambda = LambdaSequence(lmax = 1, lmin = 0.01, cardinal = 5)
)
mysem$selected

# Obtaining results in matrix format
LavaanMatrix(vect = mysem$selected[2, ], adjacency = dag)
theta # simulated is the same

}
\references{
\insertRef{RegSEM}{sharp}

\insertRef{lavaanBook}{sharp}
}
\seealso{
\code{\link{SelectionAlgo}}, \code{\link{VariableSelection}},
  \code{\link{LavaanMatrix}}

Other underlying algorithm functions: 
\code{\link{ClusteringAlgo}()},
\code{\link{PenalisedGraphical}()},
\code{\link{PenalisedRegression}()}
}
\concept{underlying algorithm functions}
