% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_plot.R
\name{plot.incremental}
\alias{plot.incremental}
\alias{IncrementalPlot}
\alias{PlotIncremental}
\title{Visualisation of incremental performance}
\usage{
\method{plot}{incremental}(
  x,
  quantiles = c(0.05, 0.95),
  ylab = "Performance",
  pch = 18,
  col = "black",
  col.axis = NULL,
  cex = 1,
  cex.lab = 1.5,
  xcex.axis = 1,
  ycex.axis = 1,
  xlas = 2,
  ylas = 1,
  sfrac = 0.005,
  ylim = NULL,
  bty = "o",
  xgrid = FALSE,
  ygrid = FALSE,
  output_data = FALSE,
  ...
)

IncrementalPlot(
  x,
  quantiles = c(0.05, 0.95),
  ylab = "Performance",
  pch = 18,
  col = "black",
  col.axis = NULL,
  cex = 1,
  cex.lab = 1.5,
  xcex.axis = 1,
  ycex.axis = 1,
  xlas = 2,
  ylas = 1,
  sfrac = 0.005,
  ylim = NULL,
  bty = "o",
  xgrid = FALSE,
  ygrid = FALSE,
  output_data = FALSE,
  ...
)

PlotIncremental(
  x,
  quantiles = c(0.05, 0.95),
  ylab = "Performance",
  pch = 18,
  col = "black",
  col.axis = NULL,
  cex = 1,
  cex.lab = 1.5,
  xcex.axis = 1,
  ycex.axis = 1,
  xlas = 2,
  ylas = 1,
  sfrac = 0.005,
  ylim = NULL,
  bty = "o",
  xgrid = FALSE,
  ygrid = FALSE,
  output_data = FALSE,
  ...
)
}
\arguments{
\item{x}{output of \code{\link{Incremental}}.}

\item{quantiles}{quantiles defining the lower and upper bounds.}

\item{ylab}{label of the y-axis.}

\item{pch}{type of point, as in \code{\link[graphics]{points}}.}

\item{col}{vector of point colours.}

\item{col.axis}{optional vector of label colours. If \code{col.axis=NULL},
the colours provided in argument \code{col} are used.}

\item{cex}{size of point.}

\item{cex.lab}{font size for labels.}

\item{xcex.axis}{size of labels along the x-axis.}

\item{ycex.axis}{size of labels along the y-axis.}

\item{xlas}{orientation of labels on the x-axis, as \code{las} in
\code{\link[graphics]{par}}.}

\item{ylas}{orientation of labels on the y-axis, as \code{las} in
\code{\link[graphics]{par}}.}

\item{sfrac}{size of the end bars, as in \code{\link[plotrix]{plotCI}}.}

\item{ylim}{displayed range along the y-axis. Only used if \code{stability}
is the output of \code{\link{BiSelection}}.}

\item{bty}{character string indicating if the box around the plot should be
drawn. Possible values include: \code{"o"} (default, the box is drawn), or
\code{"n"} (no box).}

\item{xgrid}{logical indicating if a vertical grid should be drawn. Only used
if \code{stability} is the output of \code{\link{BiSelection}}.}

\item{ygrid}{logical indicating if a horizontal grid should be drawn. Only
used if \code{stability} is the output of \code{\link{BiSelection}}.}

\item{output_data}{logical indicating if the median and quantiles should be
returned in a matrix.}

\item{...}{additional plotting arguments (see \code{\link[graphics]{par}}).}
}
\value{
A plot.
}
\description{
Represents prediction performances upon sequential inclusion of the
predictors in a logistic or Cox regression model as produced by
\code{\link{Incremental}}. The median and \code{quantiles} of the performance
metric are reported. See examples in \code{\link{Incremental}}.
}
\seealso{
\code{\link{Incremental}}
}
