% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Mtimes_SHARP.R
\name{run_Mtimes_SHARP}
\alias{run_Mtimes_SHARP}
\title{Run multiple-times SHARP for single-cell RNA data clustering}
\usage{
run_Mtimes_SHARP(scExp, ensize.K, reduced.ndim, partition.ncells, base.ncells,
  n.cores, Mtimes)
}
\arguments{
\item{scExp}{input single-cell expression matrix}

\item{ensize.K}{number of applications of random projection for ensemble}

\item{reduced.ndim}{the dimension to be reduced to}

\item{partition.ncells}{number of cells for each partition when using SHARP_large}

\item{base.ncells}{a base threshold of number of cells. When the number of cells of a dataset is smaller than this threshold, we use SHARP_small function; otherwise, we use SHARP_large.}

\item{n.cores}{number of cores to be used. The default is (n-1) cores, where n is the number of cores in your local computer or server.}

\item{Mtimes}{number of times to run SHARP}
}
\description{
This function is to run multiple times of SHARP for evaluating SHARP in clustering of single-cell RNA-Seq data
}
\examples{
enresults = run_Mtimes_SHARP(scExp, ensize.K, reduced.ndim, partition.ncells, base.ncells, n.cores, Mtimes)

}
