% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{SelectionAlgo}
\alias{SelectionAlgo}
\title{Variable selection algorithm}
\usage{
SelectionAlgo(
  xdata,
  ydata = NULL,
  Lambda,
  family = NULL,
  implementation = PenalisedRegression,
  ...
)
}
\arguments{
\item{xdata}{matrix of predictors with observations as rows and variables as
columns.}

\item{ydata}{optional vector or matrix of outcome(s). If \code{family} is set
to \code{"binomial"} or \code{"multinomial"}, \code{ydata} can be a vector
with character/numeric values, or a factor.}

\item{Lambda}{matrix of parameters controlling the level of sparsity in the
underlying feature selection algorithm specified in \code{implementation}.
With \code{implementation=PenalisedRegression}, \code{Lambda} contains
penalty parameters. If \code{Lambda=NULL},
\code{\link{LambdaGridRegression}} is used to define a relevant grid. If
\code{implementation} is not set to \code{PenalisedRegression},
\code{Lambda} must be provided.}

\item{family}{type of regression model. If
\code{implementation=PenalisedRegression}, this argument is defined as in
\code{\link[glmnet]{glmnet}}. Possible values include \code{"gaussian"}
(linear regression), \code{"binomial"} (logistic regression),
\code{"multinomial"} (multinomial regression), and \code{"cox"} (survival
analysis).}

\item{implementation}{function to use for variable selection. By default,
\code{PenalisedRegression}, based on \code{\link[glmnet]{glmnet}}, is used
for regularised regression. Other possible functions are: \code{SparsePLS},
\code{GroupPLS} and \code{SparseGroupPLS}. Alternatively, a function with
arguments \code{xdata}, \code{ydata}, \code{Lambda}, \code{family} and
\code{...}, and returning a list of two matrices named \code{selected} and
\code{beta_full} of the correct dimensions can be used.}

\item{...}{additional parameters passed to the function provided in
\code{implementation}.}
}
\value{
A list with: \item{selected}{matrix of binary selection status. Rows
  correspond to different model parameters. Columns correspond to
  predictors.} \item{beta_full}{array of model coefficients. Rows correspond
  to different model parameters. Columns correspond to predictors. Indices
  along the third dimension correspond to outcome variable(s).}
}
\description{
Runs the variable selection algorithm specified in the argument
\code{implementation} and returns matrices of model coefficients. This
function is not using stability.
}
\examples{
\dontrun{

# Data simulation
set.seed(1)
simul <- SimulateRegression(pk = 50)

# Running the LASSO
mylasso <- SelectionAlgo(
  xdata = simul$X, ydata = simul$Y,
  Lambda = c(0.1, 0.2), family = "gaussian"
)

# Simulation of additional outcomes
set.seed(2)
Y <- cbind(simul$Y, matrix(rnorm(nrow(simul$Y) * 2), ncol = 2))

# Running multivariate Gaussian LASSO
mylasso <- SelectionAlgo(
  xdata = simul$X, ydata = Y,
  Lambda = c(0.1, 0.2), family = "mgaussian"
)
str(mylasso)
}

}
\seealso{
\code{\link{VariableSelection}}

Other underlying algorithm functions: 
\code{\link{GraphicalAlgo}()},
\code{\link{PenalisedGraphical}()},
\code{\link{PenalisedRegression}()}
}
\concept{underlying algorithm functions}
