% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performances.R
\name{SelectionPerformanceGraph}
\alias{SelectionPerformanceGraph}
\title{Graph representation of selection performance}
\usage{
SelectionPerformanceGraph(
  theta,
  theta_star,
  colours = c("tomato", "forestgreen", "navy"),
  lty = c(2, 3, 1),
  plot = FALSE,
  filename = NULL,
  fileformat = "pdf",
  res = 500,
  width = 7,
  height = 7,
  units = "in",
  ...
)
}
\arguments{
\item{theta}{binary adjacency matrix.}

\item{theta_star}{true binary adjacency matrix.}

\item{colours}{vector of edge colours. The first entry of the vector defines
the colour of False Positive edges, second entry is for True Negatives and
third entry is for True Positives.}

\item{lty}{vector of line types for edges. The first entry of the vector
defines the colour of False Positive edges, second entry is for True
Negatives and third entry is for True Positives.}

\item{plot}{logical indicating if the generated graph should be plotted.}

\item{filename}{file path to saved figure. If \code{filename=NULL}, the plot
is not saved.}

\item{fileformat}{format of the saved figure. Possible values are
\code{"pdf"} or \code{"png"}. Only used if argument \code{filename} is
provided.}

\item{res}{resolution of the png figure (see \code{\link[grDevices]{png}}).
Only used if argument \code{filename} is provided and
\code{fileformat="png"}.}

\item{width}{width of the saved figure. Only used if argument \code{filename}
is provided.}

\item{height}{height of the saved figure. Only used if argument
\code{filename} is provided.}

\item{units}{units of width and height. Possible values are \code{"px"},
\code{"in"}, \code{"cm"} and \code{"mm"} (see
\code{\link[grDevices]{png}}). Only used if argument \code{filename} is
provided and \code{fileformat="png"}.}

\item{...}{additional arguments to be passed to \code{\link{Graph}}.}
}
\description{
Generates an igraph object representing the True Positive, False Positive and
False Negative edges by comparing the set of selected edges to the set of
true edges. This function only applies to graphical models and can only be
used in simulation studies (i.e. when the true model is known).
}
\examples{
\dontrun{

# Data simulation
set.seed(1)
simul <- SimulateGraphical(pk = 30)

# Stability selection
stab <- GraphicalModel(xdata = simul$data, K = 10)

# Performance graph
perfgraph <- SelectionPerformanceGraph(
  theta = Adjacency(stab),
  theta_star = simul$theta, plot = TRUE
)

# User-defined colours/shapes
perfgraph <- SelectionPerformanceGraph(
  theta = Adjacency(stab), theta_star = simul$theta, plot = TRUE,
  colours = c("forestgreen", "orange", "black"),
  node_colour = "red", node_shape = "star"
)
perfgraph <- SelectionPerformanceGraph(
  theta = Adjacency(stab), theta_star = simul$theta, plot = TRUE,
  colours = c("forestgreen", "orange", "black"), lty = c(4, 2, 3)
)

# Using and re-formatting igraph object
require(igraph)
igraph::V(perfgraph)$size <- 10
plot(perfgraph, layout = layout_with_kk(perfgraph))
}

}
\seealso{
\code{\link{GraphicalModel}}, \code{\link{Graph}}

Other selection performance functions: 
\code{\link{SelectionPerformance}()}
}
\concept{selection performance functions}
