% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{SelectionProportions}
\alias{SelectionProportions}
\title{Selection proportions}
\usage{
SelectionProportions(stability, argmax_id = NULL)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}}
or \code{\link{GraphicalModel}}.}

\item{argmax_id}{matrix of parameter IDs.
If argmax_id=NULL, selection proportions of the calibrated
stability selection model are returned.}
}
\value{
a symmetric matrix (graphical models) or vector (variable selection)
of selection proportions.
}
\description{
Extracts the selection proportions of the
(calibrated) stability selection model.
}
\examples{
## Variable selection
# Data simulation
set.seed(1)
simul=SimulateRegression(pk=50)

# Stability selection
stab=VariableSelection(xdata=simul$X, ydata=simul$Y)

# Calibrated selection proportions
prop=SelectionProportions(stab)

# User-defined parameters
myids=matrix(c(80,10), nrow=1)
stab$Lambda[myids[1],1] # corresponding penalty
stab$params$pi_list[myids[2]] # corresponding threshold
prop=SelectionProportions(stab, argmax_id=myids)

## Graphical model
# Data simulation
set.seed(1)
simul=SimulateGraphical(pk=20)

# Stability selection
stab=GraphicalModel(data=simul$data)

# Calibrated adjacency matrix
prop=SelectionProportions(stab)

# User-defined parameters
myids=matrix(c(20,10), nrow=1)
stab$Lambda[myids[1],1] # corresponding penalty
stab$params$pi_list[myids[2]] # corresponding threshold
prop=SelectionProportions(stab, argmax_id=myids)

}
