% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_output.R
\name{SelectionProportionsGraphical}
\alias{SelectionProportionsGraphical}
\title{Selection proportions (graphical model)}
\usage{
SelectionProportionsGraphical(stability, argmax_id = NULL)
}
\arguments{
\item{stability}{output of \code{\link{GraphicalModel}}.}

\item{argmax_id}{optional indices of hyper-parameters. If
\code{argmax_id=NULL}, the calibrated hyper-parameters are used.}
}
\value{
A symmetric matrix.
}
\description{
Extracts the selection proportions of the (calibrated) stability selection
model.
}
\keyword{internal}
