% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{SelectionProportionsRegression}
\alias{SelectionProportionsRegression}
\title{Selection proportions (variable selection)}
\usage{
SelectionProportionsRegression(stability, argmax_id = NULL)
}
\arguments{
\item{stability}{output of \code{\link{VariableSelection}}.}

\item{argmax_id}{optional matrix of parameter IDs. If \code{argmax_id=NULL},
the calibrated model is used.}
}
\value{
A vector of selection proportions.
}
\description{
Extracts the selection proportions of the (calibrated) stability selection
model.
}
\keyword{internal}
