% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHARP.R
\name{SHARP}
\alias{SHARP}
\title{Run SHARP for single-cell RNA data clustering}
\usage{
SHARP(scExp, ensize.K, reduced.ndim, base.ncells, partition.ncells, n.cores)
}
\arguments{
\item{scExp}{input single-cell expression matrix}

\item{ensize.K}{number of applications of random projection for ensemble}

\item{reduced.ndim}{the dimension to be reduced to}

\item{base.ncells}{a base threshold of number of cells. When the number of cells of a dataset is smaller than this threshold, we use SHARP_small function; otherwise, we use SHARP_large.}

\item{n.cores}{number of cores to be used. The default is (n-1) cores, where n is the number of cores in your local computer or server.}

\item{partition_ncells}{number of cells for each partition when using SHARP_large}
}
\description{
SHARP: \strong{S}ingle-cell RNA-Seq \strong{H}yper-fast and \strong{A}ccurate clustering via ensemble \strong{R}andom \strong{P}rojection.
}
\examples{
enresults = SHARP(scExp)

}
