% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHARP.R
\name{SHARP_small}
\alias{SHARP_small}
\title{Run SHARP for small-size (< 5000) single-cell RNA datasets}
\usage{
SHARP_small(scExp, ncells, ensize.K, reduced.ndim, hmethod, N.cluster,
  indN.cluster, minN.cluster, maxN.cluster, sil.thre, height.Ntimes, flag,
  n.cores, forview, rN.seed)
}
\arguments{
\item{scExp}{input single-cell expression matrix}

\item{ncells}{number of single cells}

\item{ensize.K}{number of applications of random projection for ensemble}

\item{reduced.ndim}{the dimension to be reduced to}
}
\description{
For small-size (< 5000) datasets, we don't need to partition the datasets into several groups. Instead, we simply use ensemble random projection and weighted ensemble meta-clustering algorithms.
}
\examples{
enresults = SHARP_small(scExp, ncells, ensize.K, reduced.ndim)

}
