% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{SimulateRegression}
\alias{SimulateRegression}
\title{Simulation of predictors and associated outcome}
\usage{
SimulateRegression(
  n = 100,
  pk = 10,
  X = NULL,
  nu_pred = 0.2,
  beta_set = c(-1, 1),
  continuous = FALSE,
  sd_pred_error = 1,
  family = "gaussian"
)
}
\arguments{
\item{n}{number of observations in the simulated data.}

\item{pk}{number of variables in the simulated data.}

\item{X}{matrix of predictors. With X=NULL, the matrix of predictors
is simulated for pk variables and n observations. If X is provided,
(a subset of) its variables will be used to simulate the outcome data.}

\item{nu_pred}{density of the set of variables to be used for the
simulation of the outcome data (i.e. number of true predictors over the
number of potential predictors).}

\item{beta_set}{related to beta coefficients used in the linear combination.
If continuous=FALSE, beta_set is the set of possible beta coefficients.
If continuous=FALSE, beta_set is the range of possible beta coefficients.}

\item{continuous}{logical indicating whether to sample from proposed
beta coefficients in "beta_set" (if continuous=FALSE) or to sample continuous
values between the minimum and maximum values in "beta_set" (if continuous=TRUE).}

\item{sd_pred_error}{residual standard deviation. Used only if family="gaussian".}

\item{family}{type of outcome. If family="gaussian", a normally distributed
outcome is simulated. If family="binomial", a binary outcome is simulated from
a Bernouilli distribution where the probability is defined as a linear combination
of (a subset of) the pk variables.}
}
\value{
A list with:
\item{X}{simulated data among which some variables were used in the linear combination
from which the outcome Y is derived. This dataset includes both signal
and noise variables.}
\item{Y}{simulated outcome Y.}
\item{proba}{true probability that the outcome is equal to 1.
Only returned if family="binomial".}
\item{logit_proba}{logit of the true probability that the outcome is equal to 1.
This is obtained as a linear combination of (a subset of) the variables in X.
Only returned if family="binomial".}
\item{theta_pred}{binary vector indicating which variables in X were used in the
linear combination to obtain the outcome Y, i.e. indicating which variables in X are
signal (if equal to 1) or noise (if equal to 0) variables.}
\item{beta}{true beta coefficients used in the linear model for simulation of the outcome Y.}
}
\description{
This function can be used to simulate (i) a matrix X of n observations
for pk normally distributed variables, and (ii) an outcome Y obtained
from the linear combination of (a subset of) the pk variables.
The outputs of this function can be used to
evaluate the ability of variable selection algorithms to identify,
among the variables in X, relevant predictors of the outcome Y.
}
\examples{
# Data simulation (continuous outcome)
simul <- SimulateRegression(n = 200, pk = 100, family = "gaussian")
plot(simul$Y_pred, simul$Y) # true linear combination vs simulated outcome
simul <- SimulateRegression(n = 200, pk = 100, family = "gaussian", sd_pred_error = 5)
plot(simul$Y_pred, simul$Y) # larger residual error

# Data simulation (binary outcome)
simul <- SimulateRegression(n = 200, pk = 100, family = "binomial")
boxplot(simul$logit_proba ~ simul$Y) # true logit probability by simulated binary outcome
}
\seealso{
Other simulation functions: 
\code{\link{SimulateGraphical}()}
}
\concept{simulation functions}
