% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{SimulateRegression}
\alias{SimulateRegression}
\title{Simulation of predictors and associated outcome}
\usage{
SimulateRegression(
  n = 100,
  pk = 10,
  X = NULL,
  theta = NULL,
  nu_pred = 0.2,
  beta_set = c(-1, 1),
  continuous = FALSE,
  sd_pred_error = 1,
  family = "gaussian"
)
}
\arguments{
\item{n}{number of observations in the simulated data.}

\item{pk}{number of variables in the simulated dataset X.}

\item{X}{matrix of predictors. With \code{X=NULL}, the matrix of predictors
is simulated for \code{pk} variables and \code{n} observations. If \code{X}
is provided, (a subset of) its variables will be used to simulate the
outcome data.}

\item{theta}{optional binary vector encoding true predictors.}

\item{nu_pred}{density of the set of variables to be used for the simulation
of the outcome data (i.e. number of true predictors over the number of
potential predictors).}

\item{beta_set}{vector defining the (range of) coefficients used in the
linear combination. If \code{continuous=FALSE}, \code{beta_set} is the set
of possible coefficients. If \code{continuous=FALSE}, \code{beta_set} is
the range of possible coefficients.}

\item{continuous}{logical indicating whether to sample coefficients from a
uniform distribution between the minimum and maximum values in
\code{beta_set} (\code{continuous=TRUE}) or from proposed values in
\code{beta_set} (\code{continuous=FALSE}).}

\item{sd_pred_error}{residual standard deviation. Used only if
\code{family="gaussian"}.}

\item{family}{type of outcome. If \code{family="gaussian"}, a normally
distributed outcome is simulated. If \code{family="binomial"}, a binary
outcome is simulated from a Bernouilli distribution where the probability
is defined as a linear combination of (a subset of) the \code{pk}
variables.}
}
\value{
A list with: \item{X}{simulated dataset. A subset of these variables
  were used in the linear combination defining the outcome Y.}
  \item{Y}{simulated outcome Y.} \item{proba}{true probability that the
  outcome is equal to 1. Only returned if \code{family="binomial"}.}
  \item{logit_proba}{logit of the true probability that the outcome is equal
  to 1. This is obtained as a linear combination of (a subset of) the
  variables in X. Only returned if \code{family="binomial"}.}
  \item{theta}{binary vector indicating which variables in X were used
  in the linear combination to obtain the outcome Y, i.e. indicating which
  variables in X are signal (if equal to 1) or noise (if equal to 0)
  variables in association with the outcome Y.} \item{beta}{true coefficients
  used in the linear model for simulation of the outcome Y.}
}
\description{
Simulate (i) a matrix X of n observations for \code{pk} normally distributed
variables, and (ii) an outcome Y obtained from a linear combination of (a
subset of) the \code{pk} variables in X. The outputs of this function can be
used to evaluate the ability of variable selection algorithms to identify,
among the variables in X, relevant predictors of the outcome Y.
}
\examples{
\dontrun{

# Data simulation (continuous outcome)
set.seed(1)
simul <- SimulateRegression(n = 200, pk = 100, family = "gaussian")
plot(simul$Y_pred, simul$Y) # true linear combination vs simulated outcome
simul <- SimulateRegression(n = 200, pk = 100, family = "gaussian", sd_pred_error = 5)
plot(simul$Y_pred, simul$Y) # larger residual error

# Outcome simulation from user-provided predictor data
set.seed(1)
X <- matrix(rnorm(n = 200), nrow = 20, ncol = 10)
simul <- SimulateRegression(X = X, family = "gaussian")

# Manually chosen set of predictors
mytheta <- c(rep(0, 8), rep(1, 2))
simul <- SimulateRegression(X = X, theta = mytheta, family = "gaussian")

# Data simulation (binary outcome)
set.seed(1)
simul <- SimulateRegression(n = 200, pk = 100, family = "binomial")
boxplot(simul$logit_proba ~ simul$Y) # true logit probability by simulated binary outcome
}
}
\seealso{
\code{\link{VariableSelection}}

Other simulation functions: 
\code{\link{SimulateAdjacency}()},
\code{\link{SimulateClustering}()},
\code{\link{SimulateGraphical}()}
}
\concept{simulation functions}
