% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Square}
\alias{Square}
\title{Adjacency from bipartite}
\usage{
Square(x)
}
\arguments{
\item{x}{matrix encoding the edges between two types of nodes (rows and
columns).}
}
\value{
A symmetric adjacency matrix encoding a bipartite graph.
}
\description{
Generates a symmetric adjacency matrix encoding a bipartite graph.
}
