% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_score.R
\name{StabilityScore}
\alias{StabilityScore}
\title{Stability score}
\usage{
StabilityScore(stab_iter, pi, K, n_cat = 3)
}
\arguments{
\item{stab_iter}{matrix or vector of selection proportions.}

\item{pi}{vector of thresholds in selection proportions.}

\item{K}{number of resampling iterations.}

\item{n_cat}{number of categories used to compute the stability score.
Possible values are 2 or 3.}
}
\value{
a vector of stability scores obtained with the different thresholds
  in selection proportions.
}
\description{
Computes the stability score from selection proportions of models with a
given penalty parameter and for different thresholds in selection
proportions.
}
\examples{
# Simulating set of selection proportions
selprop <- round(runif(n = 20), digits = 2)

# Computing stability scores for different thresholds
score <- StabilityScore(selprop, pi = c(0.6, 0.7, 0.8), K = 100)

}
\seealso{
Other stability metric functions: 
\code{\link{PFER}()}
}
\concept{stability metric functions}
